+++ tests/models/plants/test_biomasses.py
@@ -1,4 +1,4 @@
-"""Tests for the plant stoichiometry module."""
+"""Tests for the biomasses_new module."""
 
 from types import SimpleNamespace
 
@@ -7,790 +7,303 @@ import pandas as pd
 import pytest
 from numpy.testing import assert_allclose
 
+ELEMENTS = ((1, "n"), (2, "p"))
+
 
 @pytest.fixture
-def fixture_community():
-    """Provides a simple object with Community like structure."""
+def fixture_biomass_components():
+    """Provides a tuple of biomass components."""
 
-    return SimpleNamespace(
-        stem_allometry=SimpleNamespace(
-            foliage_mass=np.array([50.0, 80.0]),
-            stem_mass=np.array([100.0, 150.0]),
-            reproductive_tissue_mass=np.array([20.0, 40.0]),
-            fine_root_mass=np.array([10.0, 16.0]),
-        ),
-        cohorts=pd.DataFrame(
-            dict(
-                # TODO - revisit the ratio names and maybe automate the creation of
-                #        consistent names when loading traits to simplify the biomass
-                #        interface.
-                zeta=[0.1, 0.1],
-                sla=[2.0, 2.0],
-                p_foliage_for_reproductive_tissue=[0.5, 0.5],
-                pft_names=["shrub", "broadleaf"],
-                foliage_c_n_ratio=[5.0, 6.0],  # Why do these start differently!
-                leaf_turnover_c_n_ratio=[10.0, 12.0],
-                foliage_c_p_ratio=[5.0, 6.0],
-                leaf_turnover_c_p_ratio=[10.0, 12.0],
-                # stem_c_n_ratio=[5.0, 6.0],
-                deadwood_c_n_ratio=[5.0, 6.0],
-                # stem_c_p_ratio=[5.0, 6.0],
-                deadwood_c_p_ratio=[5.0, 6.0],
-                root_turnover_c_n_ratio=[5.0, 6.0],
-                root_turnover_c_p_ratio=[5.0, 6.0],
-                plant_reproductive_tissue_turnover_c_n_ratio=[5.0, 6.0],
-                plant_reproductive_tissue_turnover_c_p_ratio=[5.0, 6.0],
-            )
-        ),
+    cohorts = pd.DataFrame(
+        dict(
+            # TODO - revisit the ratio names and maybe automate the creation of
+            #        consistent names when loading traits to simplify the biomass
+            #        interface.
+            zeta=[0.1, 0.1],
+            sla=[2.0, 2.0],
+            p_foliage_for_reproductive_tissue=[0.5, 0.5],
+            pft_name=["shrub", "broadleaf"],
+            foliage_c_n_ratio=[5.0, 6.0],  # Why do these start differently!
+            leaf_turnover_c_n_ratio=[10.0, 12.0],
+            foliage_c_p_ratio=[5.0, 6.0],
+            leaf_turnover_c_p_ratio=[10.0, 12.0],
+            # stem_c_n_ratio=[5.0, 6.0],
+            deadwood_c_n_ratio=[5.0, 6.0],
+            # stem_c_p_ratio=[5.0, 6.0],
+            deadwood_c_p_ratio=[5.0, 6.0],
+            root_turnover_c_n_ratio=[5.0, 6.0],
+            root_turnover_c_p_ratio=[5.0, 6.0],
+            plant_reproductive_tissue_turnover_c_n_ratio=[5.0, 6.0],
+            plant_reproductive_tissue_turnover_c_p_ratio=[5.0, 6.0],
+        )
     )
 
+    stem_allometry = SimpleNamespace(
+        foliage_mass=np.array([50.0, 80.0]),
+        stem_mass=np.array([100.0, 150.0]),
+        fine_root_mass=np.array([10.0, 16.0]),
+        fruit_mass=np.array([10.0, 20.0]),
+        seed_mass=np.array([10.0, 20.0]),
+    )
 
-@pytest.fixture
-def fixture_stem_allocation():
-    """Provides a simple object with StemAllocation like structure."""
-
-    return SimpleNamespace(
+    stem_allocation = SimpleNamespace(
         foliage_turnover=np.array([30.0, 15.0]),
-        reproductive_tissue_turnover=np.array([5.0, 2.0]),
+        branch_turnover=np.array([0, 0]),
         fine_root_turnover=np.array([3.0, 1.5]),
+        fruit_turnover=np.array([2.5, 1.0]),
+        seed_turnover=np.array([2.5, 1.0]),
     )
 
-
-@pytest.fixture
-def fixture_growth_increments():
-    """Provides a simple object with GrowthIncrements like structure."""
-
-    return SimpleNamespace(
+    growth_increments = SimpleNamespace(
         delta_foliage_mass=np.array([20.0, 10.0]),
         delta_stem_mass=np.array([10.0, 5.0]),
         delta_fine_root_mass=np.array([10.0, 5.0]),
+        delta_fruit_mass=np.array([2.0, 1.0]),
+        delta_seed_mass=np.array([2.0, 1.0]),
     )
 
-
-ELEMENTS = ("N", "P")
-
-
-@pytest.fixture
-def fixture_biomasses(fixture_community):
-    """Fixture providing a Biomasses instance."""
-
-    from virtual_ecosystem.models.plants.biomasses import (
-        Biomasses,
-        Element,
-        FoliageBiomass,
-        ReproductiveBiomass,
-        RootBiomass,
-        StemBiomass,
-    )
-
-    foliage = FoliageBiomass(
-        community=fixture_community,
-        carbon_mass=fixture_community.stem_allometry.foliage_mass.copy(),
-        element_masses={
-            "N": Element(
-                name="n",
-                ideal_ratio=np.array([5.0, 6.0]),
-                actual_element_mass=np.array([5.0, 20.0]),
-                turnover_ratio=np.array([10.0, 12.0]),
-            ),
-            "P": Element(
-                name="p",
-                ideal_ratio=np.array([5.0, 6.0]),
-                actual_element_mass=np.array([5.0, 20.0]),
-                turnover_ratio=np.array([10.0, 12.0]),
-            ),
-        },
-    )
-
-    root = RootBiomass(
-        community=fixture_community,
-        carbon_mass=fixture_community.stem_allometry.fine_root_mass.copy(),
-        element_masses={
-            "N": Element(
-                name="n",
-                ideal_ratio=np.array([5.0, 7.0]),
-                actual_element_mass=np.array([10.0, 25.0]),
-                turnover_ratio=np.array(
-                    [
-                        5.0,
-                        7.0,
-                    ]
-                ),  # TODO - this is misleading. It should be current Cx_ratio
-            ),
-            "P": Element(
-                name="p",
-                ideal_ratio=np.array([5.0, 7.0]),
-                actual_element_mass=np.array([10.0, 25.0]),
-                turnover_ratio=np.array([5.0, 7.0]),
-            ),
-        },
-    )
-
-    wood = StemBiomass(
-        community=fixture_community,
-        carbon_mass=fixture_community.stem_allometry.stem_mass.copy(),
-        element_masses={
-            "N": Element(
-                name="n",
-                ideal_ratio=np.array([10.0, 8.0]),
-                actual_element_mass=np.array([5.0, 30.0]),
-                turnover_ratio=np.array([10.0, 8.0]),
-            ),
-            "P": Element(
-                name="p",
-                ideal_ratio=np.array([10.0, 8.0]),
-                actual_element_mass=np.array([5.0, 30.0]),
-                turnover_ratio=np.array([10.0, 8.0]),
-            ),  # TODO - This is sort of misleading, since the mass is zero, if we
-            # have turnover (branchfall etc) it would probably be at the current
-            # Cx_ratio
-        },
-    )
-
-    repro = ReproductiveBiomass(
-        community=fixture_community,
-        carbon_mass=fixture_community.stem_allometry.reproductive_tissue_mass.copy(),
-        element_masses={
-            "N": Element(
-                name="n",
-                ideal_ratio=np.array([4.0, 5.0]),
-                actual_element_mass=np.array([8.0, 12.0]),
-                turnover_ratio=np.array(
-                    [
-                        4.0,
-                        5.0,
-                    ]
-                ),  # TODO - again, this is misleading. It should be current Cx_ratio
-            ),
-            "P": Element(
-                name="p",
-                ideal_ratio=np.array([4.0, 5.0]),
-                actual_element_mass=np.array([8.0, 12.0]),
-                turnover_ratio=np.array([4.0, 5.0]),
-            ),
-        },
-    )
-
-    return Biomasses(
-        tissues=[foliage, repro, wood, root],
-        community=fixture_community,
-    )
-
-
-def test_Element_append():
-    """Tests the append() method of Element."""
-
-    from virtual_ecosystem.models.plants.biomasses import Element
-
-    e1 = Element(
-        name="N",
-        ideal_ratio=np.arange(10),
-        actual_element_mass=np.arange(10),
-        turnover_ratio=np.arange(10),
-    )
-
-    e2 = Element(
-        name="N",
-        ideal_ratio=np.arange(10),
-        actual_element_mass=np.arange(10),
-        turnover_ratio=np.arange(10),
-    )
-
-    e1.append(e2)
-
-    assert e1.ideal_ratio.shape == (20,)
-    assert e1.actual_element_mass.shape == (20,)
-    assert e1.turnover_ratio.shape == (20,)
-
-
-def test_BiomassTissue_append(fixture_community):
-    """Tests the shared append() method of BiomassTissueABC."""
-
-    from virtual_ecosystem.models.plants.biomasses import Element, FoliageBiomass
-
-    foliage_1 = FoliageBiomass(
-        community=fixture_community,
-        carbon_mass=fixture_community.stem_allometry.foliage_mass.copy(),
-        element_masses={
-            "N": Element(
-                name="n",
-                ideal_ratio=np.array([5.0, 6.0]),
-                actual_element_mass=np.array([5.0, 20.0]),
-                turnover_ratio=np.array([10.0, 12.0]),
-            ),
-            "P": Element(
-                name="p",
-                ideal_ratio=np.array([5.0, 6.0]),
-                actual_element_mass=np.array([5.0, 20.0]),
-                turnover_ratio=np.array([10.0, 12.0]),
-            ),
-        },
-    )
-
-    foliage_2 = FoliageBiomass(
-        community=fixture_community,
-        carbon_mass=fixture_community.stem_allometry.foliage_mass.copy(),
-        element_masses={
-            "N": Element(
-                name="n",
-                ideal_ratio=np.array([5.1, 6.1]),
-                actual_element_mass=np.array([5.1, 20.1]),
-                turnover_ratio=np.array([10.1, 12.1]),
-            ),
-            "P": Element(
-                name="p",
-                ideal_ratio=np.array([5.1, 6.1]),
-                actual_element_mass=np.array([5.1, 20.1]),
-                turnover_ratio=np.array([10.1, 12.1]),
-            ),
-        },
-    )
-
-    foliage_1.append(foliage_2)
-
-    assert_allclose(
-        foliage_1.carbon_mass, np.tile(fixture_community.stem_allometry.foliage_mass, 2)
-    )
-
-    for elem in foliage_1.element_masses.values():
-        assert_allclose(elem.ideal_ratio, np.array([5.0, 6.0, 5.1, 6.1]))
-        assert_allclose(elem.actual_element_mass, np.array([5.0, 20.0, 5.1, 20.1]))
-        assert_allclose(elem.turnover_ratio, np.array([10.0, 12.0, 10.1, 12.1]))
-
-
-def test_Biomasses_append(fixture_biomasses):
-    """Test the Biomasses append() method.
-
-    This should double the length of the mass arrays. The Element.append() and
-    BiomassTissueABC.append() methods are tested above, so this tests a subset of the
-    attributes at each level.
-    """
-
-    # Quick check that the initial values have len 2
-    for tissue in fixture_biomasses.tissues:
-        assert len(tissue.carbon_mass) == 2
-
-    # Append the biomasses fixture onto itself
-    fixture_biomasses.append(fixture_biomasses)
-
-    # Check tissue, element and surplus arrays are now len 4.
-    for tissue in fixture_biomasses.tissues:
-        assert len(tissue.carbon_mass) == 4
-
-        for elem in tissue.element_masses.values():
-            assert len(elem.actual_element_mass) == 4
-
-    for surplus in fixture_biomasses.element_surplus.values():
-        assert len(surplus) == 4
+    return cohorts, stem_allometry, stem_allocation, growth_increments
 
 
 @pytest.mark.parametrize(
-    argnames="tissue_name,mass_attribute",
+    argnames=("tissue_class"),
     argvalues=(
-        ("FoliageBiomass", "foliage_mass"),
-        ("RootBiomass", "fine_root_mass"),
-        ("StemBiomass", "stem_mass"),
-        ("ReproductiveBiomass", "reproductive_tissue_mass"),
+        pytest.param(
+            "FoliageBiomass",
+            id="foliage",
+        ),
+        pytest.param(
+            "RootBiomass",
+            id="root",
+        ),
+        pytest.param(
+            "StemBiomass",
+            id="stem",
+        ),
+        pytest.param(
+            "FruitBiomass",
+            id="fruit",
+        ),
+        pytest.param(
+            "SeedBiomass",
+            id="seed",
+        ),
     ),
 )
-def test_BiomassTissue_from_pft_default_ratios(
-    fixture_community, tissue_name, mass_attribute
+def test_BiomassTissue__init__and_methods(
+    fixture_biomass_components,
+    tissue_class,
 ):
     """Test default factory method gives a biomass tissue with the correct masses."""
 
     import virtual_ecosystem.models.plants.biomasses as biomasses
 
-    BiomassTissueClass = getattr(biomasses, tissue_name)
+    cohorts, allometry, allocation, growth_increments = fixture_biomass_components
 
-    tissue = BiomassTissueClass.from_pft_default_ratios(
-        community=fixture_community,
-        with_elements=ELEMENTS,
+    # Get the class for the parameterisation
+    BiomassTissueClass = getattr(biomasses, tissue_class)
+
+    # Get the attributes
+    mass_attr = BiomassTissueClass.mass_attr
+    turnover_mass_attr = BiomassTissueClass.turnover_mass_attr
+    growth_mass_attr = BiomassTissueClass.growth_mass_attr
+    ideal_ratio_attrs = BiomassTissueClass.ideal_ratio_attrs
+    turnover_ratio_attrs = BiomassTissueClass.turnover_ratio_attrs
+
+    # Initialise
+    tissue = BiomassTissueClass(
+        cohorts=cohorts,
+        allometry=allometry,
     )
 
     # Check carbon mass is equal to the appropriate mass attribute in the original
     # allometry.
+    assert_allclose(tissue.elemental_masses[:, 0], getattr(allometry, mass_attr))
 
-    assert_allclose(
-        tissue.carbon_mass, getattr(fixture_community.stem_allometry, mass_attribute)
-    )
-
-    # Check that the generated element masses are at their ideal ratios.
-    for ky in ELEMENTS:
+    # Check that the generated element masses are at their expected ideal ratios from
+    # the cohort traits
+    for idx, elem in ELEMENTS:
         assert_allclose(
-            tissue.element_masses[ky].actual_element_mass,
-            tissue.carbon_mass * (1 / tissue.element_masses[ky].ideal_ratio),
+            tissue.elemental_masses[:, idx],
+            tissue.elemental_masses[:, 0]
+            * (1 / cohorts[ideal_ratio_attrs.replace("ELEM", elem)]),
         )
 
+    # Deficits should all be zero.
+    assert_allclose(tissue.deficits, np.zeros_like(tissue.elemental_masses))
 
-# TODO - think about how to collapse these tests, like the one above. It might need to
-#        capture specific differences between tissues at turnover so might need a
-#        switch/case at the end, but would remove a lot of overlap.
-#        Doing this probably overlaps a lot with standardising the BiomassesABC API on
-#        fixed naming structures for element ratios.
+    # Cx_ratio property should be equal to ideals
+    assert_allclose(tissue.Cx_ratio, tissue.ideal_ratios)
 
+    # Create a tissue with non-default ratios
+    deficits = np.ones_like(tissue.elemental_masses)
+    deficits[:, 0] = 0  # no deficit in carbon
 
-def test_FoliageBiomass_functions(
-    fixture_community, fixture_stem_allocation, fixture_growth_increments
-):
-    """Test the FoliageBiomass class functions."""
-
-    from virtual_ecosystem.models.plants.biomasses import Element, FoliageBiomass
-
-    initial_element_masses = (
-        fixture_community.stem_allometry.foliage_mass.copy()
-        / fixture_community.cohorts.foliage_c_n_ratio
+    depleted_tissue = BiomassTissueClass(
+        cohorts=cohorts,
+        allometry=allometry,
+        initial_masses=tissue.elemental_masses - deficits,
     )
 
-    tissue = FoliageBiomass(
-        community=fixture_community,
-        carbon_mass=fixture_community.stem_allometry.foliage_mass.copy(),
-        element_masses={
-            "N": Element(
-                name="n",
-                ideal_ratio=fixture_community.cohorts.foliage_c_n_ratio,
-                actual_element_mass=initial_element_masses.copy(),
-                turnover_ratio=fixture_community.cohorts.leaf_turnover_c_n_ratio,
-            ),
-            "P": Element(
-                name="p",
-                ideal_ratio=fixture_community.cohorts.foliage_c_p_ratio,
-                actual_element_mass=initial_element_masses.copy(),
-                turnover_ratio=fixture_community.cohorts.leaf_turnover_c_p_ratio,
-            ),
-        },
-    )
+    # Deficits property on the tissue should equal the removed deficits.
+    assert_allclose(depleted_tissue.deficits, deficits)
 
-    tissue_defaults = FoliageBiomass.from_pft_default_ratios(
-        community=fixture_community, with_elements=ELEMENTS
-    )
+    # Replace deficits masses
+    depleted_tissue.add_elemental_masses(deficits)
 
+    # Deficits should now all be zero.
     assert_allclose(
-        tissue.as_array(with_carbon=True), tissue_defaults.as_array(with_carbon=True)
+        depleted_tissue.deficits, np.zeros_like(depleted_tissue.elemental_masses)
     )
 
-    # Check back to source: carbon mass = foliage mass
-    assert_allclose(tissue.carbon_mass, fixture_community.stem_allometry.foliage_mass)
+    # Turnover:
+    #   Calculate the values directly - this is really a bit circular but there's not an
+    #   obvious better test
+    pre_turnover_masses = tissue.elemental_masses.copy()
+    calculated_turnover = tissue.get_turnover(allocation)
 
-    # deficit = element - (C / CN)
-    expected_deficit = {
-        ky: tissue.carbon_mass / elem.ideal_ratio - elem.actual_element_mass
-        for ky, elem in tissue.element_masses.items()
-    }
-
-    calculated_deficits = tissue.deficit
-    assert calculated_deficits.keys() == expected_deficit.keys()
-    for ky in calculated_deficits:
-        assert_allclose(calculated_deficits[ky], expected_deficit[ky])
-
-    # Cx ratio = C / element
-    expected_cx = {
-        ky: tissue.carbon_mass / elem.actual_element_mass
-        for ky, elem in tissue.element_masses.items()
-    }
-
-    calculated_cx = tissue.Cx_ratio
-    assert calculated_cx.keys() == expected_cx.keys()
-    for ky in calculated_cx:
-        assert_allclose(tissue.Cx_ratio[ky], expected_cx[ky])
-
-    # turnover = turnover * (1 / turnover Cx)
-    # NOTE - TISSUE SPECIFIC DIFFERENCE - turnover tissue uses turnover Cx not
-    #        current Cx
-
-    expected_turnover = {"C": fixture_stem_allocation.foliage_turnover}
-    for ky, elem in tissue.element_masses.items():
-        expected_turnover[ky] = (
-            fixture_stem_allocation.foliage_turnover / elem.turnover_ratio
+    # Calculate from attributes
+    expected_carbon = getattr(allocation, turnover_mass_attr)
+    columns = [expected_carbon]
+    for _, elem in ELEMENTS:
+        columns.append(
+            expected_carbon
+            / cohorts[turnover_ratio_attrs.replace("ELEM", elem)].to_numpy()
         )
 
-    calculated_turnover = tissue.get_turnover(fixture_stem_allocation)
-    assert calculated_turnover.keys() == expected_turnover.keys()
-    for ky in calculated_turnover:
-        assert_allclose(calculated_turnover[ky], expected_turnover[ky])
+    # Turnover as expected
+    assert_allclose(calculated_turnover, np.stack(columns, axis=1))
 
-    # growth = ΔC / CN
-    expected_growth = {
-        ky: fixture_growth_increments.delta_foliage_mass / elem.ideal_ratio
-        for ky, elem in tissue.element_masses.items()
-    }
+    # Check no mass actually removed
+    assert_allclose(tissue.elemental_masses, pre_turnover_masses)
 
-    initial_carbon_mass = tissue.carbon_mass.copy()
+    # Growth:
+    calculated_growth = tissue.apply_growth(growth_increments)
 
-    # Run growth
-    calculated_growth = tissue.apply_growth(growth_increments=fixture_growth_increments)
-
-    # Check growth allocated
-    assert_allclose(
-        initial_carbon_mass + fixture_growth_increments.delta_foliage_mass,
-        tissue.carbon_mass,
-    )
-
-    assert calculated_growth.keys() == expected_growth.keys()
-    for ky in calculated_growth:
-        assert_allclose(calculated_growth[ky], expected_growth[ky])
-
-        assert_allclose(
-            tissue.element_masses[ky].actual_element_mass,
-            initial_element_masses + expected_growth[ky],
+    #   Again calculate the values directly - this is really a bit circular but there's
+    #   not an obvious better test
+    expected_carbon = getattr(growth_increments, growth_mass_attr)
+    columns = [expected_carbon]
+    for _, elem in ELEMENTS:
+        columns.append(
+            expected_carbon
+            / cohorts[ideal_ratio_attrs.replace("ELEM", elem)].to_numpy()
         )
 
+    # Growth as expected
+    expected_growth = np.stack(columns, axis=1)
+    assert_allclose(calculated_growth, expected_growth)
 
-def test_RootBiomass_functions(
-    fixture_community, fixture_growth_increments, fixture_stem_allocation
-):
-    """Test the RootBiomass class functions."""
-
-    from virtual_ecosystem.models.plants.biomasses import Element, RootBiomass
-
-    initial_element_masses = (
-        fixture_community.stem_allometry.fine_root_mass.copy()
-        / fixture_community.cohorts.root_turnover_c_n_ratio
-    )
-
-    tissue = RootBiomass(
-        community=fixture_community,
-        carbon_mass=fixture_community.stem_allometry.fine_root_mass.copy(),
-        element_masses={
-            "N": Element(
-                name="n",
-                ideal_ratio=fixture_community.cohorts.root_turnover_c_n_ratio,
-                actual_element_mass=initial_element_masses.copy(),
-                turnover_ratio=fixture_community.cohorts.root_turnover_c_n_ratio,
-            ),
-            "P": Element(
-                name="p",
-                ideal_ratio=fixture_community.cohorts.root_turnover_c_p_ratio,
-                actual_element_mass=initial_element_masses.copy(),
-                turnover_ratio=fixture_community.cohorts.root_turnover_c_p_ratio,
-            ),
-        },
-    )
-
-    tissue_defaults = RootBiomass.from_pft_default_ratios(
-        community=fixture_community, with_elements=ELEMENTS
-    )
-
-    assert_allclose(
-        tissue.as_array(with_carbon=True), tissue_defaults.as_array(with_carbon=True)
-    )
-
-    # carbon mass = fine root mass
-    assert_allclose(tissue.carbon_mass, fixture_community.stem_allometry.fine_root_mass)
-
-    # deficit = element - (C / CN)
-    expected_deficit = {
-        ky: tissue.carbon_mass / elem.ideal_ratio - elem.actual_element_mass
-        for ky, elem in tissue.element_masses.items()
-    }
-
-    calculated_deficits = tissue.deficit
-    assert calculated_deficits.keys() == expected_deficit.keys()
-    for ky in calculated_deficits:
-        assert_allclose(calculated_deficits[ky], expected_deficit[ky])
-
-    # Cx ratio = C / element
-    expected_cx = {
-        ky: tissue.carbon_mass / elem.actual_element_mass
-        for ky, elem in tissue.element_masses.items()
-    }
-
-    calculated_cx = tissue.Cx_ratio
-    assert calculated_cx.keys() == expected_cx.keys()
-    for ky in calculated_cx:
-        assert_allclose(tissue.Cx_ratio[ky], expected_cx[ky])
-
-    # turnover = turnover * (1 / actual Cx ratio)
-    ### NOTE TISSUE SPECIFIC DIFFERENCE - turnover tissue just has the _current_ ratios
-
-    expected_turnover = {"C": fixture_stem_allocation.fine_root_turnover}
-    for ky, elem in tissue.element_masses.items():
-        expected_turnover[ky] = (
-            fixture_stem_allocation.fine_root_turnover / expected_cx[ky]
-        )
-
-    calculated_turnover = tissue.get_turnover(fixture_stem_allocation)
-    assert calculated_turnover.keys() == expected_turnover.keys()
-    for ky in calculated_turnover:
-        assert_allclose(calculated_turnover[ky], expected_turnover[ky])
-
-    # growth = ΔC / CN
-    expected_growth = {
-        ky: fixture_growth_increments.delta_fine_root_mass / elem.ideal_ratio
-        for ky, elem in tissue.element_masses.items()
-    }
-
-    initial_carbon_mass = tissue.carbon_mass.copy()
-
-    # Run growth
-    calculated_growth = tissue.apply_growth(growth_increments=fixture_growth_increments)
-
-    # Check growth allocated
-    assert_allclose(
-        initial_carbon_mass + fixture_growth_increments.delta_fine_root_mass,
-        tissue.carbon_mass,
-    )
-
-    assert calculated_growth.keys() == expected_growth.keys()
-    for ky in calculated_growth:
-        assert_allclose(calculated_growth[ky], expected_growth[ky])
-
-        assert_allclose(
-            tissue.element_masses[ky].actual_element_mass,
-            initial_element_masses + expected_growth[ky],
-        )
-
-
-def test_ReproductiveBiomass_functions(
-    fixture_community, fixture_stem_allocation, fixture_growth_increments
-):
-    """Test the ReproductiveBiomass class functions."""
-
-    from virtual_ecosystem.models.plants.biomasses import Element, ReproductiveBiomass
-
-    initial_element_masses = (
-        fixture_community.stem_allometry.reproductive_tissue_mass.copy()
-        / fixture_community.cohorts.foliage_c_n_ratio
-    )
-
-    tissue = ReproductiveBiomass(
-        community=fixture_community,
-        carbon_mass=fixture_community.stem_allometry.reproductive_tissue_mass.copy(),
-        element_masses={
-            "N": Element(
-                name="n",
-                ideal_ratio=fixture_community.cohorts.plant_reproductive_tissue_turnover_c_n_ratio,
-                actual_element_mass=initial_element_masses.copy(),
-                turnover_ratio=fixture_community.cohorts.plant_reproductive_tissue_turnover_c_n_ratio,
-            ),
-            "P": Element(
-                name="p",
-                ideal_ratio=fixture_community.cohorts.plant_reproductive_tissue_turnover_c_p_ratio,
-                actual_element_mass=initial_element_masses.copy(),
-                turnover_ratio=fixture_community.cohorts.plant_reproductive_tissue_turnover_c_p_ratio,
-            ),
-        },
-    )
-
-    tissue_defaults = ReproductiveBiomass.from_pft_default_ratios(
-        community=fixture_community, with_elements=ELEMENTS
-    )
-
-    assert_allclose(
-        tissue.as_array(with_carbon=True), tissue_defaults.as_array(with_carbon=True)
-    )
-
-    # carbon mass = reproductive tissuemass
-    assert_allclose(
-        tissue.carbon_mass, fixture_community.stem_allometry.reproductive_tissue_mass
-    )
-
-    # deficit = element - (C / CN)
-    expected_deficit = {
-        ky: tissue.carbon_mass / elem.ideal_ratio - elem.actual_element_mass
-        for ky, elem in tissue.element_masses.items()
-    }
-
-    calculated_deficits = tissue.deficit
-    assert calculated_deficits.keys() == expected_deficit.keys()
-    for ky in calculated_deficits:
-        assert_allclose(calculated_deficits[ky], expected_deficit[ky])
-
-    # Cx ratio = C / element
-    expected_cx = {
-        ky: tissue.carbon_mass / elem.actual_element_mass
-        for ky, elem in tissue.element_masses.items()
-    }
-
-    calculated_cx = tissue.Cx_ratio
-    assert calculated_cx.keys() == expected_cx.keys()
-    for ky in calculated_cx:
-        assert_allclose(tissue.Cx_ratio[ky], expected_cx[ky])
-
-    # turnover = turnover * (1 / actual Cx ratio)
-    ### NOTE TISSUE SPECIFIC DIFFERENCE - turnover tissue just has the _current_ ratios
-
-    expected_turnover = {"C": fixture_stem_allocation.reproductive_tissue_turnover}
-    for ky, elem in tissue.element_masses.items():
-        expected_turnover[ky] = (
-            fixture_stem_allocation.reproductive_tissue_turnover / expected_cx[ky]
-        )
-
-    calculated_turnover = tissue.get_turnover(fixture_stem_allocation)
-    assert calculated_turnover.keys() == expected_turnover.keys()
-    for ky in calculated_turnover:
-        assert_allclose(calculated_turnover[ky], expected_turnover[ky])
-
-    # growth = ΔC / CN
-    carbon_growth = (
-        fixture_growth_increments.delta_foliage_mass
-        * fixture_community.cohorts.p_foliage_for_reproductive_tissue
-    )
-
-    expected_growth = {
-        ky: (carbon_growth) / elem.ideal_ratio
-        for ky, elem in tissue.element_masses.items()
-    }
-
-    initial_carbon_mass = tissue.carbon_mass.copy()
-
-    # Run growth
-    calculated_growth = tissue.apply_growth(fixture_growth_increments)
-
-    # Check growth allocated
-    assert_allclose(initial_carbon_mass + carbon_growth, tissue.carbon_mass)
-
-    assert calculated_growth.keys() == expected_growth.keys()
-    for ky in calculated_growth:
-        assert_allclose(calculated_growth[ky], expected_growth[ky])
-
-        assert_allclose(
-            tissue.element_masses[ky].actual_element_mass,
-            initial_element_masses + expected_growth[ky],
-        )
-
-
-def test_StemBiomass_functions(
-    fixture_community, fixture_stem_allocation, fixture_growth_increments
-):
-    """Test the StemBiomass class functions."""
-
-    from virtual_ecosystem.models.plants.biomasses import Element, StemBiomass
-
-    initial_element_masses = (
-        fixture_community.stem_allometry.stem_mass.copy()
-        / fixture_community.cohorts.deadwood_c_n_ratio
-    )
-
-    tissue = StemBiomass(
-        community=fixture_community,
-        carbon_mass=fixture_community.stem_allometry.stem_mass.copy(),
-        element_masses={
-            "N": Element(
-                name="n",
-                ideal_ratio=fixture_community.cohorts.deadwood_c_n_ratio,
-                actual_element_mass=initial_element_masses.copy(),
-                turnover_ratio=fixture_community.cohorts.deadwood_c_n_ratio,
-            ),
-            "P": Element(
-                name="p",
-                ideal_ratio=fixture_community.cohorts.deadwood_c_n_ratio,
-                actual_element_mass=initial_element_masses.copy(),
-                turnover_ratio=fixture_community.cohorts.deadwood_c_p_ratio,
-            ),
-        },
-    )
-
-    tissue_defaults = StemBiomass.from_pft_default_ratios(
-        community=fixture_community, with_elements=ELEMENTS
-    )
-
-    assert_allclose(
-        tissue.as_array(with_carbon=True), tissue_defaults.as_array(with_carbon=True)
-    )
-
-    # carbon mass = stem mass
-    assert_allclose(tissue.carbon_mass, fixture_community.stem_allometry.stem_mass)
-
-    # deficit = element - (C / CN)
-    expected_deficit = {
-        ky: tissue.carbon_mass / elem.ideal_ratio - elem.actual_element_mass
-        for ky, elem in tissue.element_masses.items()
-    }
-
-    calculated_deficits = tissue.deficit
-    assert calculated_deficits.keys() == expected_deficit.keys()
-    for ky in calculated_deficits:
-        assert_allclose(calculated_deficits[ky], expected_deficit[ky])
-
-    # Cx ratio = C / element
-    expected_cx = {
-        ky: tissue.carbon_mass / elem.actual_element_mass
-        for ky, elem in tissue.element_masses.items()
-    }
-
-    calculated_cx = tissue.Cx_ratio
-    assert calculated_cx.keys() == expected_cx.keys()
-    for ky in calculated_cx:
-        assert_allclose(tissue.Cx_ratio[ky], expected_cx[ky])
-
-    # turnover = zero
-    ### NOTE TISSUE SPECIFIC DIFFERENCE no turnover in stem
-    expected_turnover = {"C": np.zeros_like(fixture_community.stem_allometry.stem_mass)}
-    for ky, elem in tissue.element_masses.items():
-        expected_turnover[ky] = np.zeros_like(
-            fixture_community.stem_allometry.stem_mass
-        )
-
-    calculated_turnover = tissue.get_turnover(fixture_stem_allocation)
-
-    assert calculated_turnover.keys() == expected_turnover.keys()
-    for ky in calculated_turnover:
-        assert_allclose(calculated_turnover[ky], expected_turnover[ky])
-
-    # growth = ΔC / CN
-
-    expected_growth = {
-        ky: (fixture_growth_increments.delta_stem_mass) / elem.ideal_ratio
-        for ky, elem in tissue.element_masses.items()
-    }
-
-    initial_carbon_mass = tissue.carbon_mass.copy()
-
-    # Run growth
-    calculated_growth = tissue.apply_growth(fixture_growth_increments)
-
-    # Check growth allocated
-    assert_allclose(
-        initial_carbon_mass + fixture_growth_increments.delta_stem_mass,
-        tissue.carbon_mass,
-    )
-
-    assert calculated_growth.keys() == expected_growth.keys()
-    for ky in calculated_growth:
-        assert_allclose(calculated_growth[ky], expected_growth[ky])
-
-        assert_allclose(
-            tissue.element_masses[ky].actual_element_mass,
-            initial_element_masses + expected_growth[ky],
-        )
-
-
-def test_Biomasses_from_community(fixture_community):
-    """Test the default biomass generation."""
+    # Biomasses are now larger
+    assert_allclose(tissue.elemental_masses, pre_turnover_masses + expected_growth)
+
+    # Proportional carbon biomass, easy since only one of each PFT so 100% of carbon
+    # biomass per PFT
+    prop_carbon = tissue.get_relative_carbon_biomass_by_pft(cohorts=cohorts)
+
+    assert_allclose(prop_carbon, np.ones(2))
+
+
+@pytest.mark.parametrize(
+    argnames="tissue_class",
+    argvalues=(
+        pytest.param("FoliageBiomass", id="foliage"),
+        pytest.param("RootBiomass", id="root"),
+        pytest.param("StemBiomass", id="stem"),
+        pytest.param("FruitBiomass", id="fruit"),
+        pytest.param("SeedBiomass", id="seed"),
+    ),
+)
+def test_BiomassTissue_append(fixture_biomass_components, tissue_class):
+    """Tests the append() method of BiomassTissueABC."""
+
+    import virtual_ecosystem.models.plants.biomasses as biomasses
+
+    cohorts, allometry, _, _ = fixture_biomass_components
+
+    BiomassTissueClass = getattr(biomasses, tissue_class)
+
+    # Generate two biomass objects, one with twice the masses
+    biomasses_1 = BiomassTissueClass(cohorts=cohorts, allometry=allometry)
+    biomasses_2 = BiomassTissueClass(cohorts=cohorts, allometry=allometry)
+    biomasses_2.elemental_masses *= 2
+
+    # Append
+    biomasses_1.append(biomasses_2)
+
+    # Check shapes
+    for attr in ("elemental_masses", "ideal_ratios", "turnover_ratios"):
+        assert getattr(biomasses_1, attr).shape == (4, 3)
+
+    # Check biomass values
+    biomass_array = biomasses_1.elemental_masses
+    assert_allclose(biomass_array[2:, :] / biomass_array[:2, :], np.ones((2, 3)) * 2)
+
+    # Check cx ratios unaltered - all masses doubled in new cohorts.
+    cx_ratios = biomasses_1.Cx_ratio
+    assert_allclose(cx_ratios[2:, :], cx_ratios[:2, :])
+
+
+@pytest.fixture
+def fixture_biomasses(fixture_biomass_components):
+    """Fixture providing a Biomasses instance using the __init__ method."""
+
     from virtual_ecosystem.models.plants.biomasses import (
         Biomasses,
         FoliageBiomass,
-        ReproductiveBiomass,
+        FruitBiomass,
         RootBiomass,
+        SeedBiomass,
         StemBiomass,
     )
 
-    biomasses = Biomasses.default_init(
-        community=fixture_community,
-        with_elements=["N", "P"],
-        tissues=[FoliageBiomass, ReproductiveBiomass, StemBiomass, RootBiomass],
+    cohorts, allometry, _, _ = fixture_biomass_components
+    foliage = FoliageBiomass(cohorts=cohorts, allometry=allometry)
+    root = RootBiomass(cohorts=cohorts, allometry=allometry)
+    stem = StemBiomass(cohorts=cohorts, allometry=allometry)
+    fruit = FruitBiomass(cohorts=cohorts, allometry=allometry)
+    seed = SeedBiomass(cohorts=cohorts, allometry=allometry)
+    return Biomasses(tissues=[foliage, stem, root, fruit, seed])
+
+
+def test_Biomasses_from_cohorts(fixture_biomass_components, fixture_biomasses):
+    """Test the biomass from_cohorts class method gives the same result as direct."""
+    from virtual_ecosystem.models.plants.biomasses import (
+        Biomasses,
+        FoliageBiomass,
+        FruitBiomass,
+        RootBiomass,
+        SeedBiomass,
+        StemBiomass,
     )
 
-    # Check the biomasses are all populated correctly and the element masses are at the
-    # ideal ratios
-    for tissue_name, allom_attr in (
-        ("foliage", "foliage_mass"),
-        ("stem", "stem_mass"),
-        ("plant_reproductive_tissue", "reproductive_tissue_mass"),
-        ("root", "fine_root_mass"),
-    ):
-        tissue = biomasses.get_tissue(tissue_name)
+    cohorts, allometry, _, _ = fixture_biomass_components
 
+    biomasses = Biomasses.from_cohorts(
+        cohorts=cohorts,
+        allometry=allometry,
+        tissues=[FoliageBiomass, FruitBiomass, SeedBiomass, StemBiomass, RootBiomass],
+    )
+
+    for tissue_name in biomasses.tissue_names:
+        tissue = biomasses.get_tissue(tissue_name)
+        fixture_tissue = fixture_biomasses.get_tissue(tissue_name)
+
+        # Check the biomasses and ratios are all populated correctly
         assert_allclose(
-            tissue.carbon_mass,
-            getattr(fixture_community.stem_allometry, allom_attr),
+            tissue.elemental_masses,
+            fixture_tissue.elemental_masses,
         )
 
-        for elem in biomasses.elements:
-            element = tissue.element_masses[elem]
-            assert_allclose(
-                element.actual_element_mass, tissue.carbon_mass / element.ideal_ratio
-            )
+        assert_allclose(
+            tissue.ideal_ratios,
+            fixture_tissue.ideal_ratios,
+        )
+
+        assert_allclose(
+            tissue.turnover_ratios,
+            fixture_tissue.turnover_ratios,
+        )
 
 
 def test_total_element_mass_and_deficit(fixture_biomasses):
@@ -802,88 +315,66 @@ def test_total_element_mass_and_deficit(fixture_biomasses):
 
     for elem in fixture_biomasses.elements:
         # total element = sum across tissues
-        masses = [
-            t.element_masses[elem].actual_element_mass
-            for t in fixture_biomasses.tissues
-        ]
+        masses = [t.elemental_masses for t in fixture_biomasses.tissues]
+        assert_allclose(calculated_element_masses, np.add.reduce(masses))
 
-        assert_allclose(calculated_element_masses[elem], np.add.reduce(masses))
-
-        # total deficit = sum ideal - actual across tissues.
-        deficit = [
-            (t.carbon_mass / t.element_masses[elem].ideal_ratio)
-            - t.element_masses[elem].actual_element_mass
-            for t in fixture_biomasses.tissues
-        ]
-
-        assert_allclose(calculated_element_deficits[elem], np.add.reduce(deficit))
+        # total deficit
+        deficit = [t.deficits for t in fixture_biomasses.tissues]
+        assert_allclose(calculated_element_deficits, np.add.reduce(deficit))
 
 
 def test_apply_growth_updates_element_masses_and_surplus(
-    fixture_community,
+    fixture_biomass_components,
     fixture_biomasses,
-    fixture_stem_allocation,
-    fixture_growth_increments,
 ):
     """Test that accounting for growth updates element masses and surplus correctly."""
 
-    before = [t.as_array() for t in fixture_biomasses.tissues]
+    _, _, _, growth_increments = fixture_biomass_components
 
-    fixture_biomasses.apply_growth(growth_increments=fixture_growth_increments)
-    after = [t.as_array() for t in fixture_biomasses.tissues]
+    before = [t.elemental_masses for t in fixture_biomasses.tissues]
+    fixture_biomasses.apply_growth(growth_increments=growth_increments)
+    after = [t.elemental_masses for t in fixture_biomasses.tissues]
 
     # Each tissue should increase by element_needed_for_growth and the total surplus
     # should decrease accordingly
-    expected_surplus = np.zeros(
-        (
-            len(fixture_community.cohorts),
-            len(fixture_biomasses.elements),
-        )
-    )
+    expected_surplus = np.zeros_like(fixture_biomasses.element_surpluses)
 
     for b, a, t in zip(before, after, fixture_biomasses.tissues):
         # Test tissue increase
-        needed = t.apply_growth(growth_increments=fixture_growth_increments)
-        needed_array = np.stack(list(needed.values()))
-        expected = b + needed_array
+        needed = t.apply_growth(growth_increments=growth_increments)
+        expected = b + needed
         assert_allclose(a, expected)
 
         # Accumulate decrease in surplus
-        expected_surplus -= needed_array
+        expected_surplus -= needed
 
-    # Test accumulated surplus decrease
-    assert_allclose(
-        np.stack(list(fixture_biomasses.element_surplus.values())), expected_surplus
-    )
+    # Test accumulated surplus decrease after zeroing carbon
+    expected_surplus[:, 0] = 0
+    assert_allclose(fixture_biomasses.element_surpluses, expected_surplus)
 
 
-def test_apply_turnover(fixture_community, fixture_biomasses, fixture_stem_allocation):
+def test_apply_turnover(fixture_biomass_components, fixture_biomasses):
     """Test apply_turnover function in Biomasses."""
 
-    # Check surplus is zero going in
-    expected_surplus = {
-        el: np.zeros(len(fixture_community.cohorts))
-        for el in fixture_biomasses.elements
-    }
+    _, _, allocation, _ = fixture_biomass_components
 
-    for el in fixture_biomasses.elements:
-        assert_allclose(fixture_biomasses.element_surplus[el], expected_surplus[el])
+    # Check surplus is zero going in
+    expected_surpluses = np.zeros_like(fixture_biomasses.element_surpluses)
+    assert_allclose(fixture_biomasses.element_surpluses, expected_surpluses)
 
     # Apply the turnover
-    turnover = fixture_biomasses.apply_turnover(fixture_stem_allocation)
+    turnover = fixture_biomasses.apply_turnover(allocation)
 
     # Calculate expectations from tissues
     for t in fixture_biomasses.tissues:
-        tissue_turnover = t.get_turnover(fixture_stem_allocation)
-        for idx, el in enumerate(fixture_biomasses.elements):
-            expected_surplus[el] -= tissue_turnover[el]
-            # Check the return values match up (not currently checking C here in
-            # row zero)
-            assert_allclose(tissue_turnover[el], turnover[t.tissue_name][idx + 1])
+        tissue_turnover = t.get_turnover(allocation)
+        expected_surpluses -= tissue_turnover
+        # Check the return values match up
+        assert_allclose(tissue_turnover, turnover[t.tissue_name])
 
-    # Check surplus matches
-    for el in fixture_biomasses.elements:
-        assert_allclose(fixture_biomasses.element_surplus[el], expected_surplus[el])
+    # Check surplus matches after zeroing carbon
+    expected_surpluses[:, 0] = 0
+    assert_allclose(fixture_biomasses.element_surpluses, expected_surpluses)
 
 
 BALANCE_FOLIAGE_C = np.array([100.0, 200.0, 300.0, 400.0])
@@ -1169,7 +660,7 @@ def fixture_balance_elements_test_cases(which):
 
 @pytest.mark.parametrize("which", (0, 1, 2, 3, 4, 5, 6, 7, [0, 1, 2, 3, 4, 5, 6, 7]))
 def test_balance_elements(
-    fixture_community, fixture_balance_elements_test_cases, which
+    fixture_biomass_components, fixture_balance_elements_test_cases, which
 ):
     """Test the balancing of elements across tissues.
 
@@ -1182,7 +673,6 @@ def test_balance_elements(
 
     from virtual_ecosystem.models.plants.biomasses import (
         Biomasses,
-        Element,
         FoliageBiomass,
         StemBiomass,
     )
@@ -1197,105 +687,102 @@ def test_balance_elements(
         expected_pool,
     ) = fixture_balance_elements_test_cases
 
+    cohorts, allometry, _, _ = fixture_biomass_components
+
     n_cases = 1 if isinstance(which, int) else len(which)
 
-    # Turnover not relevant to this test
-    turnover_ratios = np.repeat(np.nan, n_cases)
+    cohorts = pd.DataFrame(
+        dict(
+            deadwood_c_n_ratio=np.tile(BALANCE_WOOD_CN, n_cases),
+            deadwood_c_p_ratio=np.tile(BALANCE_WOOD_CP, n_cases),
+            foliage_c_n_ratio=np.tile(BALANCE_FOLIAGE_CN, n_cases),
+            foliage_c_p_ratio=np.tile(BALANCE_FOLIAGE_CP, n_cases),
+            leaf_turnover_c_n_ratio=np.repeat(np.nan, n_cases * len(BALANCE_WOOD_C)),
+            leaf_turnover_c_p_ratio=np.repeat(np.nan, n_cases * len(BALANCE_WOOD_C)),
+        )
+    )
+
+    allometry = SimpleNamespace(
+        stem_mass=np.tile(BALANCE_WOOD_C, n_cases),
+        foliage_mass=np.tile(BALANCE_FOLIAGE_C, n_cases),
+    )
 
     foliage = FoliageBiomass(
-        community=fixture_community,
-        # Note that this only has two cohorts
-        # - not sure BiomassTissue will retain community as an argument.
-        carbon_mass=np.tile(BALANCE_FOLIAGE_C, n_cases),
-        element_masses={
-            "N": Element(
-                name="n",
-                ideal_ratio=np.tile(BALANCE_FOLIAGE_CN, n_cases),
-                actual_element_mass=initial_foliage[0],
-                turnover_ratio=turnover_ratios,
-            ),
-            "P": Element(
-                name="p",
-                ideal_ratio=np.tile(BALANCE_FOLIAGE_CP, n_cases),
-                actual_element_mass=initial_foliage[1],
-                turnover_ratio=turnover_ratios,
-            ),
-        },
+        cohorts=cohorts,
+        allometry=allometry,
+        initial_masses=np.concatenate(
+            [
+                allometry.foliage_mass[None, :],
+                initial_foliage,
+            ]
+        ).T,
     )
 
     wood = StemBiomass(
-        community=fixture_community,
-        carbon_mass=np.tile(BALANCE_WOOD_C, n_cases),
-        element_masses={
-            "N": Element(
-                name="n",
-                ideal_ratio=np.tile(BALANCE_WOOD_CN, n_cases),
-                actual_element_mass=initial_wood[0],
-                turnover_ratio=turnover_ratios,
-            ),
-            "P": Element(
-                name="p",
-                ideal_ratio=np.tile(BALANCE_WOOD_CP, n_cases),
-                actual_element_mass=initial_wood[1],
-                turnover_ratio=turnover_ratios,
-            ),
-        },
+        cohorts=cohorts,
+        allometry=allometry,
+        initial_masses=np.concatenate(
+            [
+                allometry.stem_mass[None, :],
+                initial_wood,
+            ]
+        ).T,
     )
 
     biomasses = Biomasses(
         tissues=[foliage, wood],
-        community=fixture_community,
+        element_surpluses=np.concatenate(
+            [
+                np.zeros_like(allometry.stem_mass)[None, :],
+                initial_pool,
+            ]
+        ).T,
     )
-
-    # Set the element pools to balance
-    biomasses.element_surplus = {
-        ky: initial_pool[idx] for idx, ky in enumerate(biomasses.elements)
-    }
 
     # Run the method.
     biomasses.balance_elements()
 
-    # Check the expectations.
-    foliage = biomasses.get_tissue("foliage").as_array()
-    assert_allclose(foliage, expected_foliage)
+    # Check the non carbon elements are as expected
+    assert_allclose(
+        biomasses.get_tissue("foliage").elemental_masses[:, 1:], expected_foliage.T
+    )
 
-    wood = biomasses.get_tissue("stem").as_array()
-    assert_allclose(wood, expected_wood)
+    assert_allclose(
+        biomasses.get_tissue("stem").elemental_masses[:, 1:], expected_wood.T
+    )
 
-    pool = np.stack(list(biomasses.element_surplus.values()))
-    assert_allclose(pool, expected_pool)
+    assert_allclose(biomasses.element_surpluses[:, 1:], expected_pool.T)
 
 
-def test_add_elemental_masses_clips_negative_value(fixture_community, caplog):
+def test_add_elemental_masses_clips_negative_value(fixture_biomass_components, caplog):
     """Tiny floating-point negatives are clipped to zero after updates."""
 
-    from virtual_ecosystem.models.plants.biomasses import Element, FoliageBiomass
+    from virtual_ecosystem.models.plants.biomasses import FoliageBiomass
+
+    cohorts, allometry, _, _ = fixture_biomass_components
+
+    initial_masses = np.stack(
+        [
+            allometry.foliage_mass.copy(),
+            np.array([1.0e-18, 20.0]),
+            np.array([5.0, 20.0]),
+        ],
+        axis=1,
+    )
 
     tissue = FoliageBiomass(
-        community=fixture_community,
-        carbon_mass=fixture_community.stem_allometry.foliage_mass.copy(),
-        element_masses={
-            "N": Element(
-                name="n",
-                ideal_ratio=np.array([5.0, 6.0]),
-                actual_element_mass=np.array([1.0e-18, 20.0]),
-                turnover_ratio=np.array([10.0, 12.0]),
-            ),
-            "P": Element(
-                name="p",
-                ideal_ratio=np.array([5.0, 6.0]),
-                actual_element_mass=np.array([5.0, 20.0]),
-                turnover_ratio=np.array([10.0, 12.0]),
-            ),
-        },
+        cohorts=cohorts,
+        allometry=allometry,
+        initial_masses=initial_masses,
     )
 
     tissue.add_elemental_masses(
-        {
-            "N": np.array([-1.1e-18, 0.0]),
-            "P": np.array([0.0, 0.0]),
-        }
+        np.stack([np.zeros(2), np.array([-1.1e-18, 0.0]), np.array([0.0, 0.0])], axis=1)
     )
 
-    assert_allclose(tissue.element_masses["N"].actual_element_mass, [0.0, 20.0])
-    assert "Clipping negative updated N biomass" in caplog.text
+    # The initial masses should be the same, except that the N mass in the first cohort
+    # has been driven down to but _not below_ zero.
+    initial_masses[0, 1] = 0
+
+    assert_allclose(tissue.elemental_masses, initial_masses)
+    assert "Clipping negative updated biomasses" in caplog.text
