+++ virtual_ecosystem/models/plants/biomasses.py
@@ -1,862 +1,380 @@
-"""The :mod:`~virtual_ecosystem.models.plants.biomasses` module contains the class
-for managing plant cohort carbon biomass along with the biomasses of all elements being
-tracked stochiometrically within a simulation. The class contains representations of the
-carbon mass and element masses within individual tissues and provides methods to balance
-stoichiometric nutrients across tissues.
-
-The theoretical tissue masses for individuals in a cohort are derived from the T Model
-predictions for the stem and the elemental masses are currently populated using the
-ideal ratio of those elements for each tissue. Growth is modelled using allocation of
-NPP from the T Model. However, the actual realised masses of different tissues can
+"""The :mod:`~virtual_ecosystem.models.plants.biomasses` module contains structures
+to track the realised stoichiometric biomasses of stems in the simulation. The
+theoretical tissue masses for individuals in a cohort are derived from the T Model
+predictions for the stem and growth and turnover are modelled using the GPP allocation
+model from the T Model. However, the actual realised masses of different tissues can
 differ from the theory due to herbivory and fruit production, so this class is used to
-track the actual carbon masses realised by individuals through the simulation.
+track the actual elemental masses realised by individuals through the simulation.
 
-The module define a base class for tissues and then currently four tissue types.
+* The :class:`BiomassTissueABC` abstract base class and subclasses implementing
+  biomasses for modelled plant tissues. The tissue biomass classes track the actual
+  masses of carbon and elemental nutrients for stems in each cohort and provide methods
+  to track changes elemental masses during tissue turnover, growth and herbivory.
 
-FoliageBiomass:
-    # Has different ratios in turnover mass
+* The :class:`Biomasses` class provides a wrapper around a set of tissues for cohorts,
+  providing methods to initialise all tissues and handle turnover and growth across
+  tissues. It also tracks the overall nutrient status of tissues and provides methods to
+  balance nutrient requirements across tissues and track nutrient deficits and
+  surpluses.
 
-    biomass: foliage_mass
-    turnover_biomass: foliage_turnover
-    growth_biomass: delta_foliage_mass
-
-    ideal_ratio: foliage_c_{elem.lower()}_ratio
-    turnover_ratio: leaf_turnover_c_{elem.lower()}_ratio
-
-    TODO - check to make sure turnover foliage doesn't get relatively _enriched_ if the
-           plant is severely nutrient depleted.
-
-ReproductiveBiomass:
-    # Same ratios; has turnover
-
-    biomass: reproductive_tissue_mass
-    turnover_biomass: reproductive_tissue_turnover
-    growth_biomass: foliage_tissue * p_foliage_for_reproductive_tissue
-
-    ideal_ratio: not defined - identical to turnover ratio
-    turnover_ratio: plant_reproductive_tissue_turnover_c_{elem.lower()}_ratio
-
-RootBiomass:
-    # Same ratios; has turnover
-
-    biomass: fine_root_mass
-    turnover_biomass: fine_root_turnover
-    growth_biomass: delta_foliage_mass * zeta * sla
-
-    ideal_ratio: not defined - identical to turnover ratio
-    turnover_ratio: root_turnover_c_{elem.lower()}_ratio
-
-StemBiomass
-    # No turnover at present, so same ratios doesn't really make sense, but if there was
-      turnover it probably would be at these ratios.
-
-    biomass: stem_mass
-    turnover_biomass: not defined, no stem turnover
-    growth_biomass: delta_stem_mass
-
-    ideal_ratio: deadwood_c_{elem.lower()}_ratio
-    turnover_ratio: not defined (because there is no turnover)
-
-
-TODO - factor out mass accessor functions (get_biomass etc) as in the lists above and
-       then make those limited functions the abstract methods so that all the logic can
-       be shared in base class methods.
-
-TODO - make this all run on 2D CNP arrays (remove Element?) Have single arrays giving
-       biomasses and ideal and turnover ratios.
-
+Both classes provide ``append`` methods to add new cohorts into the biomass
+representation.
 """  # noqa: D205
 
 from __future__ import annotations
 
-from abc import ABC, abstractmethod
-from dataclasses import dataclass, field
+from abc import ABC
 from typing import ClassVar
 
 import numpy as np
 from numpy.typing import NDArray
-from pyrealm.demography.core import ToDataFrameMixin
-from pyrealm.demography.tmodel import GrowthIncrements, StemAllocation
+from pyrealm.demography.cohorts import Cohorts
+from pyrealm.demography.tmodel import GrowthIncrements, StemAllocation, StemAllometry
 from xarray import DataArray
 
 from virtual_ecosystem.core.logger import LOGGER
-from virtual_ecosystem.models.plants.communities import Community
 
 
-@dataclass
-class Element:
-    """Stochiometric elemental masses for cohorts in a community."""
+def partition_reproductive_tissue_mass(
+    cohorts: Cohorts, mass: NDArray[np.floating]
+) -> tuple[NDArray[np.floating], NDArray[np.floating]]:
+    """Splits a reproductive tissue mass into fruit and seed components.
 
-    name: str
-    """The element name."""
-    ideal_ratio: NDArray[np.floating]
-    """The ideal ratio of the element for the tissue type."""
-    actual_element_mass: NDArray[np.floating]
-    """The actual mass of the element for the tissue type."""
-    turnover_ratio: NDArray[np.floating]
-    """What to do with this on non-reclaiming tissues."""
+    This is used to split reproductive tissue mass, turnover or delta into fruit and
+    seed components.
 
-    def append(self, other: Element) -> None:
-        """Appends new data representing new cohorts onto an element instance."""
-        self.ideal_ratio = np.append(self.ideal_ratio, other.ideal_ratio)
-        self.actual_element_mass = np.append(
-            self.actual_element_mass, other.actual_element_mass
-        )
-        self.turnover_ratio = np.append(self.turnover_ratio, other.turnover_ratio)
+    Args:
+        cohorts: A cohort dataframe providing trait data.
+        mass: An array providing a reproductive tissue mass for each cohort.
+    """
+
+    fruit_mass = mass * cohorts["fruit_flesh_fraction"].to_numpy()
+    return fruit_mass, mass - fruit_mass
 
 
-@dataclass
 class BiomassTissueABC(ABC):
-    """A dataclass to hold tissue stoichiometry biomasses for a set of plant cohorts.
+    """Stoichiometry tissue biomasses for a set of plant cohorts.
 
-    This class holds the current quantity of a given element (generally N or P) for a
-    specific plant tissue type (generally foliage, wood, roots or reproductive tissue).
-    The class also holds the ideal ratio of the element for that tissue type. They hold
-    an entry for each cohort in the data class.
+    This base class holds the current elemental masses for a specific plant tissue (e.g.
+    foliage, stem, fine roots) for a set of cohorts, along with the tissue-specicic
+    ideal and turnover nutrient C/N ratios for each cohort.
+
+    * The elemental biomasses are held as a numpy array with a column for each element
+      and a row for each cohort. Carbon masses are always in the first column, followed
+      by the nutrient elements given in the the ``elements`` class attribute. The
+      initial biomasses are set using the allometric carbon mass of the tissue and the
+      ideal ratios for the tissue, unless these are overridden by providing an array of
+      ``initial_masses``.
+
+    * The ideal and turnover ratios are also held as numpy arrays, with the same shape
+      as the biomasses and are expressed as C/x ratios. The first column of these arrays
+      is therefore C/C and is always equal to one.
+
+    The class provides methods to add and remove biomasses through turnover, growth and
+    herboivory and to add new cohorts to the biomass representation. It also has
+    properties to access current realised C/x ratios and nutrient deficits.
+
+    Subclasses of this abstract base class provide implementations for specific tissues
+    and are defined simply by setting class attributes that give the names of tissue
+    specific mass and ratio attributes from model objects:
+
+    * ``tissue_name``: a tissue label
+    * ``mass_attr``: the tissue mass attribute name in StemAllometry objects.
+    * ``turnover_mass_attr``: the tissue turnover mass attribute name in StemAllocation
+      objects.
+    * ``growth_mass_attr``: The tissue growth mass attribute name in GrowthIncrements
+      objects.
+    * ``ideal_ratio_attrs``: A string that can be used to match the tissue ideal ratio
+      attributes defined in the flora and included in Cohorts objects
+    * ``turnover_ratio_attrs``: As above but matching tissue turnover ratio attributes.
+      These will typically point to the same attribute as the ideal ratio - the tissue
+      does not resorb nutrients during turnover - but can identify different ratios
+      where resorption does occur.
+
+    The last two attributes must be defined using the format ``stem_c_ELEM_ratio``,
+    where ELEM can be be replaced with the lower case element letter to match the
+    specific elemental ratio traits (e.g. ``stem_c_n_ratio``, ``stem_c_p_ratio``)
+
+    Args:
+        cohorts: A dataframe of cohort data.
+        allometry: A StemAllometry object for the cohorts.
+        initial_masses: An optional initial set of elemental biomasses.
     """
 
     tissue_name: ClassVar[str]
     """A tissue name for derived classes."""
-    community: Community
-    """The community object that the tissue is associated with."""
+    mass_attr: ClassVar[str]
+    """The tissue mass attribute name in StemAllometry objects"""
+    turnover_mass_attr: ClassVar[str]
+    """The tissue turnover mass attribute name in StemAllocation objects"""
+    growth_mass_attr: ClassVar[str]
+    """The tissue growth mass attribute name in GrowthIncrements objects"""
+    ideal_ratio_attrs: ClassVar[str]
+    """The form of the tissue ideal ratio attributes defined in the flora and available 
+    from Cohorts objects"""
+    turnover_ratio_attrs: ClassVar[str]
+    """The form of the tissue turnover ratio attributes defined in the flora and
+    available from Cohorts objects"""
 
-    carbon_mass: NDArray[np.floating]
-    """An 1D array of tissue carbon mass for each stem in cohorts in the community."""
-    element_masses: dict[str, Element]
-    """A dictionary of Elements containing nutrient masses each stem in cohorts in the
-    community."""
+    elements: ClassVar[tuple[str, ...]] = ("N", "P")
+    """A tuple giving the nutrient elements included in the biomasses."""
 
-    @classmethod
-    @abstractmethod
-    def from_pft_default_ratios(
-        cls,
-        community: Community,
-        with_elements: list[str],
+    def __init__(
+        self,
+        cohorts: Cohorts,
+        allometry: StemAllometry,
+        initial_masses: NDArray[np.floating] | None = None,
     ):
         """Create a default instance of Tissue based on the PFT traits."""
 
+        # Need to use copy to avoid the biomass and allometry masses refer to the same
+        # object!
+
+        self.elemental_masses: NDArray[np.floating]
+        """An 2D array of tissue elemental masses stems in cohorts."""
+
+        self.turnover_ratios: NDArray[np.floating]
+        """Elemental ratios for tissue turnover. These may be identical to ideal ratios 
+        if the tissue does have nutrient resorption on turnover tissue."""
+
+        self.ideal_ratios: NDArray[np.floating]
+        """Ideal elemental ratios for the tissue."""
+
+        self.set_ideal_ratios(cohorts=cohorts)
+        self.set_turnover_ratios(cohorts=cohorts)
+
+        if initial_masses is not None:
+            if not initial_masses.shape == (len(cohorts), len(self.elements) + 1):
+                raise ValueError(
+                    "Shape of provided elemental masses not compatible with number "
+                    "of cohorts and tracked nutrients."
+                )
+            self.elemental_masses = initial_masses
+            return
+
+        # Otherwise set from ideal ratios
+        carbon_mass = getattr(allometry, self.mass_attr)
+        self.elemental_masses = carbon_mass[:, None] / self.ideal_ratios
+
+    def get_ratios(self, cohorts, ratio_attrs) -> NDArray[np.floating]:
+        """Get an array of the ideal or turnover ratios for a tissue.
+
+        Args:
+            cohorts: The cohort data providing ratios
+            ratio_attrs: The name form for the different elemental ratio attributes.
+        """
+        return np.stack(
+            [
+                np.ones(len(cohorts)),
+                *[
+                    cohorts[ratio_attrs.replace("ELEM", elem.lower())].to_numpy()
+                    for elem in self.elements
+                ],
+            ],
+            axis=1,
+        )
+
+    def set_ideal_ratios(self, cohorts: Cohorts) -> None:
+        """Method to set ideal nutrient ratios."""
+        self.ideal_ratios = self.get_ratios(
+            cohorts=cohorts, ratio_attrs=self.ideal_ratio_attrs
+        )
+
+    def set_turnover_ratios(self, cohorts: Cohorts) -> None:
+        """Abstract method to set turnover nutrient ratios."""
+        self.turnover_ratios = self.get_ratios(
+            cohorts=cohorts, ratio_attrs=self.turnover_ratio_attrs
+        )
+
     @property
-    def deficit(self) -> dict[str, NDArray[np.floating]]:
-        """Calculate the element deficit (ideal mass - actual mass) for the tissue.
+    def deficits(self) -> NDArray[np.floating]:
+        """Calculate the elemental deficits (ideal mass - actual mass) for the tissue.
 
         Returns:
-            The element deficit for the specified tissue.
+            The tissue deficits for the tissue.
         """
-        return {
-            ky: (self.carbon_mass / elem.ideal_ratio) - elem.actual_element_mass
-            for ky, elem in self.element_masses.items()
-        }
+        return (
+            self.elemental_masses[:, [0]] / self.ideal_ratios
+        ) - self.elemental_masses
 
     @property
-    def get_elemental_masses(self) -> dict[str, NDArray[np.floating]]:
-        """Get the current element masses for the tissue.
+    def Cx_ratio(self) -> NDArray[np.floating]:
+        """Get current carbon to element ratios for the tissue type.
+
+        This has to handle cases where a tissue has no biomass at all or no actual
+        elemental mass, which would otherwise generate NaN ratios (C/0). It explicitly
+        sets these cases to infinity.
+
+        .. TODO:
+
+            Not currently used in the PlantsModel and anyway returned something odd.
+            Updated to give current actual Cx ratios, but maybe can delete
 
         Returns:
-            The element deficit for the specified tissue.
+            The carbon to element ratios for the tissue.
         """
-        return {
-            ky: elem.actual_element_mass for ky, elem in self.element_masses.items()
-        }
 
-    def add_elemental_masses(self, masses: dict[str, NDArray[np.floating]]) -> None:
+        return np.where(
+            self.elemental_masses == 0,
+            np.inf,
+            self.elemental_masses[:, [0]] / self.elemental_masses,
+        )
+
+    def add_elemental_masses(self, masses: NDArray[np.floating]) -> None:
         """Add elemental masses to the tissues."""
 
-        try:
-            for ky, elem in self.element_masses.items():
-                updated_masses = elem.actual_element_mass + masses[ky]
-                # Clip any negative result from the update and log the values clipped.
-                negative_updated = updated_masses < 0.0
-                if np.any(negative_updated):
-                    LOGGER.warning(
-                        f"Clipping negative updated {ky} biomass in "
-                        f"{self.tissue_name}: {updated_masses[negative_updated]}"
-                    )
-                    updated_masses = np.clip(updated_masses, 0.0, None)
-                elem.actual_element_mass = updated_masses
-        except KeyError:
-            raise ValueError("add_elemental_masses missing required element.")
-        except ValueError:
+        if not self.elemental_masses.shape == masses.shape:
             raise ValueError("Error adding elements mass - incompatible shapes.")
 
-    def get_relative_carbon_biomass_by_pft(self) -> NDArray[np.floating]:
+        updated_masses = self.elemental_masses + masses
+        # Clip any negative result from the update and log the values clipped.
+        negative_updated = updated_masses < 0.0
+        if np.any(negative_updated):
+            LOGGER.warning(
+                f"Clipping negative updated biomasses in "
+                f"{self.tissue_name}: {updated_masses[negative_updated]}"
+            )
+        self.elemental_masses = np.clip(updated_masses, 0.0, None)
+
+    def append(self, other: BiomassTissueABC):
+        """Add new tissue data representing new cohorts."""
+
+        for attr in ("elemental_masses", "ideal_ratios", "turnover_ratios"):
+            # Append the elemental masses and ratios from the incoming instance
+            # onto the existing instance.
+            setattr(
+                self, attr, np.concatenate([getattr(self, attr), getattr(other, attr)])
+            )
+
+    def get_turnover(self, allocation: StemAllocation) -> NDArray[np.floating]:
+        """Calculate the biomass lost to turnover for foliage tissue.
+
+        This method calculates the tissue turnover from the GPP allocation model and
+        returns an array giving the biomasses lost to turnover. This does not alter the
+        biomasses within the tissue as turnover biomass is replaced by GPP allocation.
+
+        Returns:
+            The element quantity lost to turnover for foliage tissue.
+        """
+        carbon_mass = getattr(allocation, self.turnover_mass_attr)
+
+        return carbon_mass[:, None] / self.turnover_ratios
+
+    def apply_growth(self, growth_increments: GrowthIncrements) -> NDArray[np.floating]:
+        """Apply biomasses increases given growth increments.
+
+        This method adjusts the carbon biomass following the growth increments from the
+        allocation model and increases nutrient biomasses following the ideal ratios.
+        The method returns the biomass increases, so that subsequent nutrient balancing
+        can account for deficits and excesses across tissues.
+
+        Returns:
+            The increases in element quantities needed to support growth at the ideal
+            ratio for the tissue.
+        """
+        carbon_mass = getattr(growth_increments, self.growth_mass_attr)
+        nutrient_ideal_ratio_increase = carbon_mass[:, None] / self.ideal_ratios
+        self.add_elemental_masses(nutrient_ideal_ratio_increase)
+
+        return nutrient_ideal_ratio_increase
+
+    def get_relative_carbon_biomass_by_pft(
+        self, cohorts: Cohorts
+    ) -> NDArray[np.floating]:
         """Get the proportional carbon biomass of each cohort within PFTs for a tissue.
 
         This is used to distribute herbivory - which happens at the PFT level - back
         down to individual cohorts, assuming that herbivory is distributed between
         cohorts of the same PFT in proportion to the available biomass.
 
+        TODO: This will need to nest by cell id if we drop communities, since herbivory
+              is applied at the cell level.
+
         Args:
-            tissue_type: The type of tissue to retrieve (e.g., 'foliage', 'wood').
+            cohorts: The cohort data for the biomasses.
 
         Returns:
             An one-dimensional array with length equal to the number of cohorts giving
             the proportional carbon biomass of that cohort within the PFT.
-
         """
 
-        total_pft_carbon_biomass = np.zeros_like(self.carbon_mass)
+        carbon_mass = self.elemental_masses[:, 0]
+        total_pft_carbon_biomass = np.zeros_like(carbon_mass)
 
         # Use boolean indexing to collate the total PFT biomass for each cohort
-        # NOTE - this relies on the community being updated by reference when
-        #        recruitment happens. If this changes then the match of the number of
-        #        columns to the PFTs needs to be maintained some other way.
-        for pft in set(self.community.cohorts["pft_name"]):
+        for pft in set(cohorts["pft_name"]):
             # boolean index along carbon_mass array
-            in_pft = self.community.cohorts["pft_name"].to_numpy() == pft
-            # aggregate masses across cohorts in the PFT and assign total.
-            total_pft_carbon_biomass[in_pft] = self.carbon_mass[in_pft].sum()
+            in_pft = cohorts["pft_name"].to_numpy() == pft
+            # aggregate carbon masses across cohorts in the PFT and assign total.
+            total_pft_carbon_biomass[in_pft] = carbon_mass[in_pft].sum()
 
-        return self.carbon_mass / total_pft_carbon_biomass
+        return carbon_mass / total_pft_carbon_biomass
 
     def apply_herbivory(self, herbivory_array: DataArray):
-        """Remove biomass from a tissue to account for herbiivory.
+        """Remove biomass from a tissue to account for herbivory.
 
         The input is expected to be a DataArray with a pft dimension matching the number
         of cohorts and then an element dimension containing C and then each element.
-
-        NOTE - if this class moves to an all array representation of biomasses, then it
-               we should be able just to subtract the incoming array from the current
-               element masses. Note that np.array - xr.DataArray returns an xr.DataArray
-               so need to reduce to numpy.
         """
-        self.carbon_mass -= herbivory_array.sel(element="C").to_numpy()
-        for elem_name, elem in self.element_masses.items():
-            elem.actual_element_mass -= herbivory_array.sel(
-                element=elem_name
-            ).to_numpy()
-
-    @property
-    def Cx_ratio(self) -> dict[str, NDArray[np.floating]]:
-        """Get the carbon to element ratio for the tissue type.
-
-        If there is no elemental mass (and possible also no carbon mass) for any
-        cohorts in the tissue then the function return ``np.inf`` for those values.
-
-        Returns:
-            The carbon to element ratio for the specified tissue.
-        """
-        ratios = {}
-
-        for ky, elem in self.element_masses.items():
-            # Use np.divide and where here to avoid triggering warnings on zero divides
-            ratios[ky] = np.divide(
-                self.carbon_mass,
-                elem.actual_element_mass,
-                where=elem.actual_element_mass > 0,
-                out=np.full_like(self.carbon_mass, np.inf),
-            )
-
-        return ratios
-
-    def as_array(
-        self, deficit: bool = False, with_carbon: bool = False
-    ) -> NDArray[np.floating]:
-        """Utility method to return tissue masses as an array.
-
-        TODO: The internals of this class may switch over to array based, this is a
-              placeholder API for that change.
-
-        Args:
-            deficit: Return the deficit masses not the actual masses.
-            with_carbon: Should carbon mass be included in the array.
-        """
-
-        if deficit:
-            elemental_masses = list(self.deficit.values())
-        else:
-            elemental_masses = [
-                elem.actual_element_mass for elem in self.element_masses.values()
-            ]
-        if with_carbon:
-            return np.stack(
-                [self.carbon_mass, *elemental_masses],
-            )
-
-        return np.stack(elemental_masses)
-
-    @abstractmethod
-    def apply_growth(
-        self, growth_increments: GrowthIncrements
-    ) -> dict[str, NDArray[np.floating]]:
-        """Increase tissue biomasses following growth allocation for the tissue.
-
-        This method should adjust the carbon biomass following the growth increments
-        from the allocation model and similarly increase nutrient biomasses following
-        the ideal ratios. The method must then also return the nutrient biomass
-        increases, so that subsequent nutrient balancing can account for deficits and
-        excesses within the stem.
-        """
-
-    @abstractmethod
-    def get_turnover(
-        self, allocation: StemAllocation
-    ) -> dict[str, NDArray[np.floating]]:
-        """Calculate the elemental losses to turnover for the tissue type.
-
-        This method should return dictionary keyed by elements, including carbon. The
-        values should be an array giving the per stem biomass within each cohort that is
-        lost to turnover given the allocation.
-        """
-
-    def append(self, other: BiomassTissueABC):
-        """Add new tissue data representing new cohorts."""
-
-        # TODO? Checking for consistent elements
-
-        # Append the carbon mass from the incoming instance
-        self.carbon_mass = np.append(self.carbon_mass, other.carbon_mass)
-
-        # Append the element masses from the incoming instance
-        for elem_name, elem_instance in self.element_masses.items():
-            elem_instance.append(other.element_masses[elem_name])
+        self.elemental_masses -= herbivory_array.to_numpy()
 
 
-@dataclass
 class FoliageBiomass(BiomassTissueABC):
-    """A class to hold foliage stoichiometry data for a set of plant cohorts."""
+    """Foliage tissue stoichiometry biomasses for plant cohorts."""
 
     tissue_name = "foliage"
-
-    @classmethod
-    def from_pft_default_ratios(
-        cls,
-        community: Community,
-        with_elements: list[str],
-    ):
-        """Create a default instance of FoliageBiomass based on the PFT traits."""
-
-        element_masses: dict[str, Element] = {}
-
-        # Need to use copy to avoid the biomass and allometry masses refer to the same
-        # object!
-        carbon_mass = pyrealm_handling(community.stem_allometry.foliage_mass.copy())
-        for elem in with_elements:
-            ideal_ratio = community.cohorts[
-                f"foliage_c_{elem.lower()}_ratio"
-            ].to_numpy()
-
-            turnover_ratio = community.cohorts[
-                f"leaf_turnover_c_{elem.lower()}_ratio"
-            ].to_numpy()
-
-            element_masses[elem] = Element(
-                name=elem,
-                ideal_ratio=ideal_ratio,
-                actual_element_mass=carbon_mass / ideal_ratio,
-                turnover_ratio=turnover_ratio,
-            )
-
-        return cls(
-            carbon_mass=carbon_mass,
-            community=community,
-            element_masses=element_masses,
-        )
-
-    def apply_growth(
-        self, growth_increments: GrowthIncrements
-    ) -> dict[str, NDArray[np.floating]]:
-        """Increase the biomasses of foliage tissue given the allocation model.
-
-        Returns:
-            The increases in element quantities needed to support growth at the ideal
-            ratio for the tissue.
-        """
-        carbon_mass = pyrealm_handling(growth_increments.delta_foliage_mass)
-
-        self.carbon_mass += carbon_mass
-
-        nutrient_ideal_ratio_increase = {
-            ky: carbon_mass * (1 / elem.ideal_ratio)
-            for ky, elem in self.element_masses.items()
-        }
-
-        self.add_elemental_masses(nutrient_ideal_ratio_increase)
-
-        return nutrient_ideal_ratio_increase
-
-    def get_turnover(
-        self, allocation: StemAllocation
-    ) -> dict[str, NDArray[np.floating]]:
-        """Calculate the element mass lost to turnover for foliage tissue.
-
-        Returns:
-            The element quantity lost to turnover for foliage tissue.
-        """
-        carbon_mass = pyrealm_handling(allocation.foliage_turnover)
-        elemental_turnovers = {
-            ky: (carbon_mass * (1 / elem.turnover_ratio))
-            for ky, elem in self.element_masses.items()
-        }
-
-        return {"C": carbon_mass, **elemental_turnovers}
+    mass_attr = "foliage_mass"
+    turnover_mass_attr = "foliage_turnover"
+    growth_mass_attr = "delta_foliage_mass"
+    ideal_ratio_attrs = "foliage_c_ELEM_ratio"
+    turnover_ratio_attrs = "leaf_turnover_c_ELEM_ratio"
 
 
-@dataclass
-class ReproductiveBiomass(BiomassTissueABC):
-    """Holds reproductive tissue stoichiometry data for a set of plant cohorts."""
-
-    tissue_name = "plant_reproductive_tissue"
-
-    @classmethod
-    def from_pft_default_ratios(
-        cls,
-        community: Community,
-        with_elements: list[str],
-    ):
-        """Create a default instance of FoliageBiomass based on the PFT traits."""
-
-        element_masses: dict[str, Element] = {}
-
-        # Use copy to avoid maintaining a reference to the allometry
-        carbon_mass = pyrealm_handling(
-            community.stem_allometry.reproductive_tissue_mass.copy()
-        )
-
-        for elem in with_elements:
-            # Turnover ratio is identical to ideal ratio
-            ideal_ratio = community.cohorts[
-                f"plant_reproductive_tissue_turnover_c_{elem.lower()}_ratio"
-            ]
-            turnover_ratio = ideal_ratio
-
-            element_masses[elem] = Element(
-                name=elem,
-                ideal_ratio=ideal_ratio,
-                actual_element_mass=carbon_mass / ideal_ratio,
-                turnover_ratio=turnover_ratio,
-            )
-
-        return cls(
-            carbon_mass=carbon_mass,
-            community=community,
-            element_masses=element_masses,
-        )
-
-    def apply_growth(
-        self, growth_increments: GrowthIncrements
-    ) -> dict[str, NDArray[np.floating]]:
-        """Increase the biomasses of reproductive tissue given the allocation model.
-
-        Returns:
-            The increases in element quantities needed to support growth at the ideal
-            ratio for the tissue.
-        """
-
-        carbon_increase = pyrealm_handling(
-            growth_increments.delta_foliage_mass
-            * self.community.cohorts["p_foliage_for_reproductive_tissue"].to_numpy()
-        )
-        self.carbon_mass += carbon_increase
-
-        nutrient_ideal_ratio_increase = {
-            ky: (carbon_increase * (1 / elem.ideal_ratio))
-            for ky, elem in self.element_masses.items()
-        }
-
-        self.add_elemental_masses(nutrient_ideal_ratio_increase)
-
-        return nutrient_ideal_ratio_increase
-
-    def get_turnover(
-        self, allocation: StemAllocation
-    ) -> dict[str, NDArray[np.floating]]:
-        """Calculate the element mass lost to turnover for reproductive tissue.
-
-        Returns:
-            The element quantity lost to turnover for reproductive tissue.
-        """
-
-        # TODO: Caching locally to avoid calling the property constructor for each
-        # element  - maybe this should a cached property?
-
-        cx_ratios = self.Cx_ratio
-        carbon_mass = pyrealm_handling(allocation.reproductive_tissue_turnover)
-        elemental_turnovers = {
-            ky: carbon_mass * (1 / cx_ratios[ky])
-            for ky, elem in self.element_masses.items()
-        }
-
-        return {"C": carbon_mass, **elemental_turnovers}
-
-
-@dataclass
 class FruitBiomass(BiomassTissueABC):
-    """Holds fruit tissue stoichiometry data for a set of plant cohorts."""
+    """Fruit tissue stoichiometry biomasses for plant cohorts."""
 
     tissue_name = "fruit"
-
-    @classmethod
-    def from_pft_default_ratios(
-        cls,
-        community: Community,
-        with_elements: list[str],
-    ):
-        """Create a default instance of FoliageBiomass based on the PFT traits."""
-
-        element_masses: dict[str, Element] = {}
-
-        # Get the proportion of reproductive tissue allocated to fruit
-        fruit_fraction = community.cohorts["fruit_flesh_fraction"].to_numpy()
-
-        # Multiplication here avoids the need to copy() the array to avoid the reference
-        # back to the allometry
-        carbon_mass = np.array(
-            pyrealm_handling(
-                community.stem_allometry.reproductive_tissue_mass * fruit_fraction
-            )
-        )
-
-        for elem in with_elements:
-            ideal_ratio = community.cohorts[
-                f"plant_reproductive_tissue_turnover_c_{elem.lower()}_ratio"
-            ]
-            # Turnover ratio is identical to ideal ratio
-            turnover_ratio = ideal_ratio
-
-            element_masses[elem] = Element(
-                name=elem,
-                ideal_ratio=ideal_ratio,
-                actual_element_mass=carbon_mass / ideal_ratio,
-                turnover_ratio=turnover_ratio,
-            )
-
-        return cls(
-            carbon_mass=carbon_mass,
-            community=community,
-            element_masses=element_masses,
-        )
-
-    def apply_growth(
-        self, growth_increments: GrowthIncrements
-    ) -> dict[str, NDArray[np.floating]]:
-        """Increase the biomasses of reproductive tissue given the allocation model.
-
-        Returns:
-            The increases in element quantities needed to support growth at the ideal
-            ratio for the tissue.
-        """
-
-        fruit_fraction = self.community.cohorts["fruit_flesh_fraction"].to_numpy()
-
-        carbon_increase = pyrealm_handling(
-            growth_increments.delta_foliage_mass
-            * self.community.cohorts["p_foliage_for_reproductive_tissue"].to_numpy()
-            * fruit_fraction
-        )
-        self.carbon_mass += carbon_increase
-
-        nutrient_ideal_ratio_increase = {
-            ky: (carbon_increase * (1 / elem.ideal_ratio))
-            for ky, elem in self.element_masses.items()
-        }
-
-        self.add_elemental_masses(nutrient_ideal_ratio_increase)
-
-        return nutrient_ideal_ratio_increase
-
-    def get_turnover(
-        self, allocation: StemAllocation
-    ) -> dict[str, NDArray[np.floating]]:
-        """Calculate the element mass lost to turnover for reproductive tissue.
-
-        Returns:
-            The element quantity lost to turnover for reproductive tissue.
-        """
-
-        # Get the proportion of reproductive tissue allocated to fruit
-        carbon_turnover = (
-            pyrealm_handling(allocation.reproductive_tissue_turnover)
-            * self.community.cohorts["fruit_flesh_fraction"].to_numpy()
-        )
-
-        # TODO: Caching locally to avoid calling the property constructor for each
-        # element  - maybe this should a cached property?
-        cx_ratios = self.Cx_ratio
-
-        elemental_turnovers = {
-            ky: (carbon_turnover * (1 / cx_ratios[ky]))
-            for ky, elem in self.element_masses.items()
-        }
-
-        return {"C": carbon_turnover, **elemental_turnovers}
+    mass_attr = "fruit_mass"
+    turnover_mass_attr = "fruit_turnover"
+    growth_mass_attr = "delta_fruit_mass"
+    ideal_ratio_attrs = "plant_reproductive_tissue_turnover_c_ELEM_ratio"
+    turnover_ratio_attrs = "plant_reproductive_tissue_turnover_c_ELEM_ratio"
 
 
-@dataclass
 class SeedBiomass(BiomassTissueABC):
-    """Holds fruit tissue stoichiometry data for a set of plant cohorts."""
+    """Seed tissue stoichiometry biomasses for plant cohorts."""
 
     tissue_name = "seed"
-
-    @classmethod
-    def from_pft_default_ratios(
-        cls,
-        community: Community,
-        with_elements: list[str],
-    ):
-        """Create a default instance of FoliageBiomass based on the PFT traits."""
-
-        element_masses: dict[str, Element] = {}
-
-        # Get the proportion of reproductive tissue allocated to seed
-        # - Multiplication here avoids the need to copy() the array to avoid the
-        #   reference back to the allometry
-        carbon_mass = np.array(
-            pyrealm_handling(
-                community.stem_allometry.reproductive_tissue_mass
-                * (1 - community.cohorts["fruit_flesh_fraction"].to_numpy())
-            )
-        )
-
-        for elem in with_elements:
-            # Turnover ratio is identical to ideal ratio
-            ideal_ratio = community.cohorts[
-                f"plant_reproductive_tissue_turnover_c_{elem.lower()}_ratio"
-            ]
-            turnover_ratio = ideal_ratio
-
-            element_masses[elem] = Element(
-                name=elem,
-                ideal_ratio=ideal_ratio,
-                actual_element_mass=carbon_mass / ideal_ratio,
-                turnover_ratio=turnover_ratio,
-            )
-
-        return cls(
-            carbon_mass=carbon_mass,
-            community=community,
-            element_masses=element_masses,
-        )
-
-    def apply_growth(
-        self, growth_increments: GrowthIncrements
-    ) -> dict[str, NDArray[np.floating]]:
-        """Increase the biomasses of reproductive tissue given the allocation model.
-
-        Returns:
-            The increases in element quantities needed to support growth at the ideal
-            ratio for the tissue.
-        """
-
-        # Get the proportion of reproductive tissue allocated to seed
-        carbon_increase = pyrealm_handling(
-            growth_increments.delta_foliage_mass
-            * self.community.cohorts["p_foliage_for_reproductive_tissue"].to_numpy()
-            * self.community.cohorts["fruit_flesh_fraction"].to_numpy()
-        )
-        self.carbon_mass += carbon_increase
-
-        nutrient_ideal_ratio_increase = {
-            ky: (carbon_increase * (1 / elem.ideal_ratio))
-            for ky, elem in self.element_masses.items()
-        }
-
-        self.add_elemental_masses(nutrient_ideal_ratio_increase)
-
-        return nutrient_ideal_ratio_increase
-
-    def get_turnover(
-        self, allocation: StemAllocation
-    ) -> dict[str, NDArray[np.floating]]:
-        """Calculate the element mass lost to turnover for reproductive tissue.
-
-        Returns:
-            The element quantity lost to turnover for reproductive tissue.
-        """
-
-        # TODO: Caching locally to avoid calling the property constructor for each
-        # element  - maybe this should a cached property?
-
-        cx_ratios = self.Cx_ratio
-
-        # Get the proportion of reproductive tissue allocated to seed
-        carbon_turnover = pyrealm_handling(allocation.reproductive_tissue_turnover) * (
-            1 - self.community.cohorts["fruit_flesh_fraction"].to_numpy()
-        )
-
-        elemental_turnovers = {
-            ky: (carbon_turnover * (1 / cx_ratios[ky]))
-            for ky, elem in self.element_masses.items()
-        }
-
-        return {"C": carbon_turnover, **elemental_turnovers}
+    mass_attr = "seed_mass"
+    turnover_mass_attr = "seed_turnover"
+    growth_mass_attr = "delta_seed_mass"
+    ideal_ratio_attrs = "plant_reproductive_tissue_turnover_c_ELEM_ratio"
+    turnover_ratio_attrs = "plant_reproductive_tissue_turnover_c_ELEM_ratio"
 
 
-@dataclass
 class StemBiomass(BiomassTissueABC):
-    """A class to hold stem stoichiometry data for a set of plant cohorts."""
+    """Stem tissue stoichiometry biomasses for plant cohorts."""
 
     tissue_name = "stem"
-
-    @classmethod
-    def from_pft_default_ratios(
-        cls,
-        community: Community,
-        with_elements: list[str],
-    ):
-        """Create a default instance of WoodBiomass based on the PFT traits."""
-
-        element_masses: dict[str, Element] = {}
-
-        # Use copy to avoid maintaining a reference to the allometry
-        carbon_mass = pyrealm_handling(community.stem_allometry.stem_mass.copy())
-
-        for elem in with_elements:
-            ideal_ratio = community.cohorts[f"deadwood_c_{elem.lower()}_ratio"]
-            turnover_ratio = np.zeros_like(ideal_ratio)
-
-            element_masses[elem] = Element(
-                name=elem,
-                ideal_ratio=ideal_ratio,
-                actual_element_mass=carbon_mass / ideal_ratio,
-                turnover_ratio=turnover_ratio,
-            )
-
-        return cls(
-            carbon_mass=carbon_mass,
-            community=community,
-            element_masses=element_masses,
-        )
-
-    def apply_growth(
-        self, growth_increments: GrowthIncrements
-    ) -> dict[str, NDArray[np.floating]]:
-        """Increase the biomasses of woody tissue given the allocation model.
-
-        Returns:
-            The increases in element quantities needed to support growth at the ideal
-            ratio for the tissue.
-        """
-        self.carbon_mass += pyrealm_handling(growth_increments.delta_stem_mass)
-
-        nutrient_ideal_ratio_increase = {
-            ky: pyrealm_handling(growth_increments.delta_stem_mass)
-            * (1 / elem.ideal_ratio)
-            for ky, elem in self.element_masses.items()
-        }
-
-        self.add_elemental_masses(nutrient_ideal_ratio_increase)
-
-        return nutrient_ideal_ratio_increase
-
-    def get_turnover(
-        self, allocation: StemAllocation
-    ) -> dict[str, NDArray[np.floating]]:
-        """Calculate the element mass lost to turnover for foliage tissue.
-
-        Returns:
-            The element quantity lost to turnover for foliage tissue.
-        """
-        elemental_turnovers = {
-            ky: np.zeros_like(self.carbon_mass)
-            for ky, elem in self.element_masses.items()
-        }
-
-        return {"C": np.zeros_like(self.carbon_mass), **elemental_turnovers}
+    mass_attr = "stem_mass"
+    turnover_mass_attr = "branch_turnover"
+    growth_mass_attr = "delta_stem_mass"
+    ideal_ratio_attrs = "deadwood_c_ELEM_ratio"
+    turnover_ratio_attrs = "deadwood_c_ELEM_ratio"
 
 
-@dataclass
 class RootBiomass(BiomassTissueABC):
-    """A class to hold root stoichiometry data for a set of plant cohorts."""
+    """Fine root tissue stoichiometry biomasses for plant cohorts."""
 
     tissue_name = "root"
-
-    @classmethod
-    def from_pft_default_ratios(
-        cls,
-        community: Community,
-        with_elements: list[str],
-    ):
-        """Create a default instance of FoliageBiomass based on the PFT traits."""
-
-        element_masses: dict[str, Element] = {}
-
-        for elem in with_elements:
-            ideal_ratio = community.cohorts[f"root_turnover_c_{elem.lower()}_ratio"]
-            turnover_ratio = np.ones_like(ideal_ratio)
-
-            element_masses[elem] = Element(
-                name=elem,
-                ideal_ratio=ideal_ratio,
-                actual_element_mass=community.stem_allometry.fine_root_mass
-                / ideal_ratio,
-                turnover_ratio=turnover_ratio,
-            )
-
-        return cls(
-            carbon_mass=community.stem_allometry.fine_root_mass,
-            community=community,
-            element_masses=element_masses,
-        )
-
-    def apply_growth(
-        self, growth_increments: GrowthIncrements
-    ) -> dict[str, NDArray[np.floating]]:
-        """Increase the biomasses of root tissue given the allocation model.
-
-        Returns:
-            The increases in element quantities needed to support growth at the ideal
-            ratio for the tissue.
-        """
-
-        carbon_increase = pyrealm_handling(growth_increments.delta_fine_root_mass)
-
-        self.carbon_mass += carbon_increase
-
-        nutrient_ideal_ratio_increase = {
-            ky: (carbon_increase * (1 / elem.ideal_ratio))
-            for ky, elem in self.element_masses.items()
-        }
-
-        self.add_elemental_masses(nutrient_ideal_ratio_increase)
-
-        return nutrient_ideal_ratio_increase
-
-    def get_turnover(
-        self, allocation: StemAllocation
-    ) -> dict[str, NDArray[np.floating]]:
-        """Calculate the element mass lost to turnover for foliage tissue.
-
-        Returns:
-            The element quantity lost to turnover for foliage tissue.
-        """
-
-        cx_ratios = self.Cx_ratio
-        carbon_mass = pyrealm_handling(allocation.fine_root_turnover)
-
-        elemental_turnovers = {
-            ky: (carbon_mass * (1 / cx_ratios[ky]))
-            for ky, elem in self.element_masses.items()
-        }
-
-        return {"C": carbon_mass, **elemental_turnovers}
+    mass_attr = "fine_root_mass"
+    turnover_mass_attr = "fine_root_turnover"
+    growth_mass_attr = "delta_fine_root_mass"
+    ideal_ratio_attrs = "root_turnover_c_ELEM_ratio"
+    turnover_ratio_attrs = "root_turnover_c_ELEM_ratio"
 
 
-def pyrealm_handling(arr: NDArray):
-    """Handle pyrealm dimensions.
-
-    The pyrealm package in version 2.0 and 3.0 returns 2D arrays of cohort data that
-    have length 1 in the second dimension: np.array([[1.5, 1.6]]). Those can be
-    collapsed using squeeze() _but_ this breaks on communities with a single cohort
-    because squeeze is greedy and collapses np.array([[1.5]]) to the zero-dimension case
-    np.array(1.5).
-
-    You can specify _which_ axis to squeeze (e.g. squeeze(1)) but the test data uses 1D
-    arrays as inputs, and it seems likely that pyrealm should change to stop doing this
-    as a default. So for the moment this function is used to simplify 2D to 1D.
-
-    Args:
-        arr: A numpy array.
-    """
-
-    if arr.ndim > 1:
-        # Specifically suppress pyrealm arrays with length 1 in axis 1.
-        return arr.squeeze(0)
-
-    return arr
-
-
-@dataclass
-class Biomasses(ToDataFrameMixin):
+class Biomasses:  # TODO - ToDataFrameMixin? Some kind of export method
     """A class holding biomasses for a set of plant cohorts and tissues.
 
     This class holds the current ratios across tissue type for a community object, which
@@ -867,43 +385,60 @@ class Biomasses(ToDataFrameMixin):
     required.
     """
 
-    tissues: list[BiomassTissueABC]
-    """Tissues for the associated cohorts."""
-    community: Community
-    """The community object that the stoichiometry is associated with."""
-    element_surpluses: dict[str, NDArray[np.floating]] = field(init=False)
-    """The surplus of the element per cohort."""
-    tissue_names: list[str] = field(init=False)
-    """A list giving the name of each tissue."""
-    elements: tuple[str, ...] = field(init=False)
-    """A list of the elements recorded in each tissue."""
-
-    # Note: these are hard-coded and must be updated if the simulation
-    # uses different biomass classes.
+    # NOTE: these are hard-coded and must be updated if the simulation
+    #       uses different biomass classes.
+    # TODO: Might also be redundant if the ToDataFramMixin approach isn't going to be
+    #       used, which it might well not be - can just concat the tissue arrays into a
+    #       data frame.
     _array_attrs: ClassVar[tuple[str, ...]] = (
-        "biomass_foliage_carbon_mass",
-        "biomass_foliage_n_actual_element_mass",
-        "biomass_foliage_p_actual_element_mass",
-        "biomass_fruit_carbon_mass",
-        "biomass_fruit_n_actual_element_mass",
-        "biomass_fruit_p_actual_element_mass",
-        "biomass_seed_carbon_mass",
-        "biomass_seed_n_actual_element_mass",
-        "biomass_seed_p_actual_element_mass",
-        "biomass_stem_carbon_mass",
-        "biomass_stem_n_actual_element_mass",
-        "biomass_stem_p_actual_element_mass",
-        "biomass_root_carbon_mass",
-        "biomass_root_n_actual_element_mass",
-        "biomass_root_p_actual_element_mass",
+        "foliage_c_biomass",
+        "foliage_n_biomass",
+        "foliage_p_biomass",
+        "fruit_c_biomass",
+        "fruit_n_biomass",
+        "fruit_p_biomass",
+        "seed_c_biomass",
+        "seed_n_biomass",
+        "seed_p_biomass",
+        "stem_c_biomass",
+        "stem_n_biomass",
+        "stem_p_biomass",
+        "root_c_biomass",
+        "root_n_biomass",
+        "root_p_biomass",
     )
     """Array attribute names for all biomass tissue and element data."""
 
-    def __post_init__(self) -> None:
-        """Initialize the element surplus for each cohort."""
+    def __init__(
+        self,
+        tissues: list[BiomassTissueABC],
+        element_surpluses: NDArray[np.floating] | None = None,
+    ) -> None:
+
+        # TODO - do we actually need this constructor as opposed to just using the
+        #        from_cohorts method as __init__?
+
+        self.tissues: list[BiomassTissueABC] = tissues
+        """Tissues for the associated cohorts."""
+        self.element_surpluses: NDArray[np.floating]
+        """The surplus of the element per cohort."""
+        self.tissue_names: list[str]
+        """A list giving the name of each tissue."""
+        self.elements: tuple[str, ...]
+        """A list of the elements recorded in each tissue."""
+
+        # Check the elemental mass shapes are consistent
+        element_shape_set = {t.elemental_masses.shape for t in self.tissues}
+
+        if len(element_shape_set) > 1:
+            raise ValueError(
+                f"Tissues passed to StemBiomasses have different elemental "
+                f"mass array shapes: {element_shape_set}"
+            )
+        element_shapes = element_shape_set.pop()
 
         # Check the elements being used
-        elements = {tuple(tissue.element_masses.keys()) for tissue in self.tissues}
+        elements = {t.elements for t in self.tissues}
 
         if len(elements) > 1:
             raise ValueError(
@@ -915,77 +450,72 @@ class Biomasses(ToDataFrameMixin):
         self.elements = elements.pop()
 
         # Populate the whole individual elemental surpluses
-        # TODO - should this populate from the tissues themselves. When this is from
-        # default_init then these _will_ be zeros, but that isn't true of direct use of
-        # the constructor. Is there ever a case we'd use the __init__ though?
-
-        self.element_surplus = {
-            elem: np.zeros(len(self.community.cohorts)) for elem in self.elements
-        }
+        if element_surpluses is not None:
+            if element_surpluses.shape != element_shapes:
+                raise ValueError(
+                    "Provided element surpluses have a different shape to the tissue "
+                    "biomass arrays."
+                )
+            self.element_surpluses = element_surpluses
+        else:
+            self.element_surpluses = np.zeros(element_shapes)
 
     @classmethod
-    def default_init(
+    def from_cohorts(
         cls,
-        community: Community,
-        with_elements: list[str],
+        cohorts: Cohorts,
+        allometry: StemAllometry,
         tissues: list[type[BiomassTissueABC]],
     ):
-        """Create an instance of StemStoichiometry from the PFT stoichiometry ratios.
+        """Create a Biomasses instance from cohort data using the ideal element ratios.
 
         Args:
-            community: The community object that the stoichiometry is associated with.
-            extra_pft_traits: Additional traits specific to the plant functional type.
-            with_elements: The name of the elements to be used in the biomass
-                representation.
+            cohorts: A data frame of cohorts providing trait data.
+            allometry: The allometry of the cohorts.
             tissues: A list of tissue models to be used.
 
         Returns:
-            An instance of StemBiomasses with default tissues for the community.
+            An instance of Biomasses with default element ratios for the cohorts.
         """
 
         # Generate the default tissues
         default_tissues: list[BiomassTissueABC] = [
-            tissue.from_pft_default_ratios(
-                community=community,
-                with_elements=with_elements,
+            tissue(
+                cohorts=cohorts,
+                allometry=allometry,
             )
             for tissue in tissues
         ]
 
-        return cls(
-            tissues=default_tissues,
-            community=community,
-        )
+        return cls(tissues=default_tissues)
 
     @property
-    def total_element_masses(self) -> dict[str, NDArray[np.floating]]:
+    def total_element_masses(self) -> NDArray[np.floating]:
         """Calculate the total element mass for each cohort.
 
         Returns:
-            The total element mass for each cohort.
+            The total element mass across tissues for each cohort.
         """
-        masses = [t.get_elemental_masses for t in self.tissues]
-        return {ky: np.add.reduce([m[ky] for m in masses]) for ky in self.elements}
+        return np.add.reduce([t.elemental_masses for t in self.tissues])
 
     @property
-    def tissue_deficit(self) -> dict[str, NDArray[np.floating]]:
+    def tissue_deficit(self) -> NDArray[np.floating]:
         """Calculate the total element deficits for each cohort.
 
         Returns:
             The element deficit for all cohorts.
         """
-        deficits = [t.deficit for t in self.tissues]
-        return {ky: np.add.reduce([d[ky] for d in deficits]) for ky in self.elements}
+
+        return np.add.reduce([t.deficits for t in self.tissues])
 
     def _adjust_surpluses(
-        self, masses: dict[str, NDArray[np.floating]], increase: bool = True
+        self, masses=NDArray[np.floating], increase: bool = True
     ) -> None:
         """Adjust the element surpluses in the biomass object."""
-        for elem in self.elements:
-            if increase:
-                self.element_surplus[elem] += masses[elem]
-            else:
-                self.element_surplus[elem] -= masses[elem]
+        if increase:
+            self.element_surpluses += masses
+        else:
+            self.element_surpluses -= masses
 
     def apply_growth(self, growth_increments: GrowthIncrements) -> None:
         """Distribute the carbon allocated to growth and required nutrients to tissues.
@@ -1008,7 +538,8 @@ class Biomasses(ToDataFrameMixin):
             # add that mass at ideal ratios.
             needed = tissue.apply_growth(growth_increments=growth_increments)
             # Record the nutrients biomasses at ideal ratios allocated to the tissue in
-            # the whole stem balance.
+            # the whole stem balance, after zeroing carbon which is allocated from NPP.
+            needed[:, 0] = 0
             self._adjust_surpluses(needed, increase=False)
 
     def apply_turnover(
@@ -1034,28 +565,22 @@ class Biomasses(ToDataFrameMixin):
           accumulated deficits and gains to those pools at the end of the allocation
           process.
 
-        TODO: This needs to account for herbivory - proportional to herbivory loss?
-
         Returns:
             A dictionary by tissue of turnover biomass arrays
         """
 
         # Get turnover by tissue
         turnover_by_tissue = {
-            tissue.tissue_name: np.stack(list(tissue.get_turnover(allocation).values()))
-            for tissue in self.tissues
+            t.tissue_name: t.get_turnover(allocation) for t in self.tissues
         }
-
-        # TODO - this next bit is clunky but moving the whole system over to arrays not
-        #        dicts internally will probably fix it.
 
         # Accumulate the tissue specific turnovers into a single whole stem turnover
         total_turnover = np.add.reduce(list(turnover_by_tissue.values()))
 
-        # This total is a CNP by cohorts array and needs to be split back to per element
-        # dictionary of arrays.
-        total_turnover_by_element = dict(zip(["C", *self.elements], total_turnover))
-        self._adjust_surpluses(total_turnover_by_element, increase=False)
+        # Remove carbon from losses - replaced by turnover allocation - and update
+        # surpluses
+        total_turnover[:, 0] = 0
+        self._adjust_surpluses(total_turnover, increase=False)
 
         return turnover_by_tissue
 
@@ -1073,13 +598,8 @@ class Biomasses(ToDataFrameMixin):
 
         # Get 3D arrays of elements/cohorts/tissue for actual element masses and
         # per tissue element deficits
-        tissue_element_masses = np.stack([t.as_array() for t in self.tissues])
-        tissue_element_deficits = np.stack(
-            [t.as_array(deficit=True) for t in self.tissues]
-        )
-
-        # Get a 2D array of elements/cohort from the individual-level element pool
-        stem_pools = np.stack(list(self.element_surplus.values()))
+        tissue_element_masses = np.stack([t.elemental_masses for t in self.tissues])
+        tissue_element_deficits = np.stack([t.deficits for t in self.tissues])
 
         # Calculate the redistribution of pool deficits (negative values) to tissues
         # weighted by the relative elemental mass for each tissue.
@@ -1090,7 +610,7 @@ class Biomasses(ToDataFrameMixin):
         # more of that element removed since they have none.
 
         tissue_total_element_masses = tissue_element_masses.sum(axis=0)
-        pool_deficits_to_tissues = stem_pools * np.divide(
+        pool_deficits_to_tissues = self.element_surpluses * np.divide(
             tissue_element_masses,
             tissue_total_element_masses,
             out=np.zeros_like(tissue_element_masses),
@@ -1111,7 +631,7 @@ class Biomasses(ToDataFrameMixin):
             where=stem_deficits != 0,
         )
 
-        pool_surpluses_to_tissues = stem_pools * tissue_relative_deficits
+        pool_surpluses_to_tissues = self.element_surpluses * tissue_relative_deficits
 
         # Constrain element changes
         # - don't fill tissue deficits beyond the ideal ratio.
@@ -1126,18 +646,17 @@ class Biomasses(ToDataFrameMixin):
 
         # Combine the two redistribution paths to give deficits and surpluses
         pool_to_tissue = np.where(
-            stem_pools < 0, pool_deficits_to_tissues, pool_surpluses_to_tissues
+            self.element_surpluses < 0,
+            pool_deficits_to_tissues,
+            pool_surpluses_to_tissues,
         )
 
         # Allocate resulting masses back to tissues
         for tissue, to_tissue in zip(self.tissues, pool_to_tissue):
-            tissue.add_elemental_masses(
-                {ky: mass for ky, mass in zip(self.elements, to_tissue)}
-            )
+            tissue.add_elemental_masses(to_tissue)
 
         # Remove masses allocated to tissues from pool.
-        for elem, from_pool in zip(self.elements, pool_to_tissue.sum(axis=0)):
-            self.element_surplus[elem] -= from_pool
+        self.element_surpluses -= pool_to_tissue.sum(axis=0)
 
     def get_tissue(self, tissue_type: str) -> BiomassTissueABC:
         """Get the tissue model for a specific tissue type.
@@ -1162,7 +681,9 @@ class Biomasses(ToDataFrameMixin):
         for tissue_name in self.tissue_names:
             self.get_tissue(tissue_name).append(other.get_tissue(tissue_name))
 
-        for elem in self.elements:
-            self.element_surplus[elem] = np.append(
-                self.element_surplus[elem], other.element_surplus[elem]
-            )
+        self.element_surpluses = np.concatenate(
+            [
+                self.element_surpluses,
+                other.element_surpluses,
+            ]
+        )
